### 2025-11-11 #729 @maRce10
suwo: Access Nature Media Repositories
Submitting Author Name: Marcelo Araya Salas
Submitting Author Github Handle: @maRce10
Repository:  https://github.com/maRce10/suwo
Version submitted:
Submission type: Standard

Version accepted: TBD

---

```
Package: suwo
Title: Access Nature Media Repositories Through R
Version: 0.1.0
Maintainer: Marcelo Araya-Salas <marcelo.araya@ucr.ac.cr>
Description: Streamline searching/downloading of nature media files (e.g. audios, photos) from online repositories. The package offers functions for,obtaining media metadata from online repositories, downloading associated media files ad updating data sets with new records.
    checkmate,
    cli,
    pbapply,
    viridis,
    httr,
    kableExtra,
    tools,
    RecordLinkage,
    jpeg,
    fs
    covr,
    leaflet,
    graphics
    c(person("Marcelo", "Araya-Salas",
      email = "marcelo.araya@ucr.ac.cr",
      comment = c(ORCID = "0000-0003-3594-619X")),
      person("Jorge", "Elizondo-Calvo",
      role = c("aut")),
      person("Alejandro", "Rico-Guevara",
      role = c("aut"),
      email = "colibri@uw.edu",
      comment = c(ORCID = "0000-0003-4067-5312")))

```

It downloads metadata and media files from nature media repositories

Nature media files are increasingly used in diverse fields ranging from ecology and evolutionary biology
(e.g. trait evolution) to wildlife monitoring and conservation (training species detection models).

The package warbleR (which I maintain) offers a tool to download media files from a single repository. The package wikiaves (https://github.com/athospd/wikiaves) does a similar thing (but does not seem to be working). suwo includes functions for 5 repositories (including those from the other packages) and also offers a single framework to combine and curate data from different sources.

Yes.

Code coverage is already 80% (https://app.codecov.io/gh/maRce10/suwo/tree/main) for branch main (pkgcheck seems to be picking up the (deleted) "master" branch results)

Yes, but not right now.
