+++ R/find_duplicates.R
@@ -43,13 +43,15 @@
 #' different platforms. This can also occur as some repositories automatically
 #' share data with other repositories, particularly with GBIF.
 #' @examples
-#' if (interactive()){
 #' # get metadata from 2 repos
 #' gb <- query_gbif(species = "Turdus rufiventris", format =  "sound")
-#' xc <- query_xenocanto(species = "Turdus rufiventris")
+#' inat <- query_inaturalist(species = "Turdus rufiventris",
+#'   format = "sound")
 #'
+#' # run if queries didnt fail
+#'  if (!is.null(gb) && !is.null(inat)) {
 #' # combine metadata
-#' merged_metadata <- merge_metadata(xc, gb)
+#' merged_metadata <- merge_metadata(inat, gb)
 #'
 #' # find duplicates
 #' label_dup_metadata <- find_duplicates(metadata = merged_metadata)
