+++ R/merge_metadata.R
@@ -14,13 +14,14 @@
 #' suwo query functions (e.g., `query_wikiaves()`, `query_xenocanto()`, etc.)
 #' with `raw_data = FALSE`.
 #' @examples
-#' if (interactive()){
-#'  # get metadata from 2 repos
-#'   wa <- query_wikiaves(species = 'Glaucis dohrnii', format =  "sound")
-#'   xc <- query_xenocanto(species = 'Glaucis dohrnii')
+#' # get metadata from 2 repos
+#' wa <- query_wikiaves(species = "Glaucis dohrnii", format =  "sound")
+#' gb <- query_gbif(species = "Glaucis dohrnii", format = "sound")
 #'
-#'   # combine metadata
-#'   merged_mt <- merge_metadata(wa, xc)
+#' # run if queries didnt fail
+#'  if (!is.null(wa) && !is.null(gb)) {
+#'  # combine metadata
+#'  merged_mt <- merge_metadata(wa, gb)
 #' }
 #'
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})

[…]
