### 2025-11-13 10ea52a
Update CITATION.cff
