+++ reference/download_media.md
@@ -103,12 +103,14 @@ Marcelo Araya-Salas (<marcelo.araya@ucr.ac.cr>)
 ## Examples
 
 ``` r
-if (interactive()){
-  phae_anth <- query_xenocanto(species = 'Phaethornis anthophilus',
-  all_data = FALSE)
+  h_peckii <- query_inaturalist(species = 'Hydnellum peckii',
+  format = "image")
 
+# run if query didnt fail
+ if (!is.null(h_peckii)) {
   # donwload the first to files
-  phae_anth_downl <- download_media(metadata = phae_anth[1:2, ],
+  phae_anth_downl <- download_media(metadata = h_peckii[1:2, ],
   path = tempdir())
 }
+#> ✔ All files were downloaded successfully 🌈
 ```
