### 2025-11-13 a5cbd5d
fix CRAN errors 2
