+++ R/update_metadata.R
@@ -32,18 +32,18 @@
 #' some old entries are not returned in the new query they are still retained.
 #' @examples
 #' # query metadata
-#' wa <- query_wikiaves(species = 'Glaucis dohrnii', format =  "sound")
+#' a_gioiosa <- query_gbif(species = "Amanita gioiosa", format =  "image")
 #'
 #' # run if query didnt fail
-#'  if (!is.null(wa)) {
+#'  if (!is.null(a_gioiosa)) {
 #' # remove last 3 rows to test update_metadata
-#' sub_wa <- wa[1:(nrow(wa)- 3), ]
+#' sub_a_gioiosa <- a_gioiosa[1:(nrow(a_gioiosa)- 3), ]
 #'
 #' # update
-#' up_wa <- update_metadata(metadata = sub_wa)
+#' up_a_gioiosa <- update_metadata(metadata = sub_a_gioiosa)
 #'
 #' # check number of rows is the same
-#' nrow(up_wa) == nrow(wa)
+#' # nrow(up_a_gioiosa) == nrow(a_gioiosa)
 #' }
 #'
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
@@ -156,6 +156,14 @@ update_metadata <-
         verbose = verbose,
         pb = pb
       )
+    }
+
+    # stop gracefully if query returned NULL
+    if (is.null(query_output_new)) {
+      if (verbose) {
+        .message("No new entries found", "failure", suffix =  "\n")
+      }
+      return(invisible(NULL))
     }
 
     # Find duplicates
