+++ testing/testing-checking suwo functions.Rmd
@@ -449,6 +449,7 @@ sim_annotation <- function(n = 10, n.sound.files = 1){
 ```{r check, eval = FALSE}
 
 #run document twice
+Sys.setenv(XENO_CANTO_API_KEY = getPass::getPass())
 devtools::document(".")
 
 devtools::check(document = TRUE, run_dont_test = TRUE, vignettes = FALSE, manual = TRUE)

[…]
