+++ reference/download_media.md
@@ -103,13 +103,13 @@ Marcelo Araya-Salas (<marcelo.araya@ucr.ac.cr>)
 ## Examples
 
 ``` r
-  h_peckii <- query_inaturalist(species = 'Hydnellum peckii',
+  a_zambiana <- query_inaturalist(species = "Amanita zambiana",
   format = "image")
 
 # run if query didnt fail
- if (!is.null(h_peckii)) {
+ if (!is.null(a_zambiana)) {
   # donwload the first to files
-  phae_anth_downl <- download_media(metadata = h_peckii[1:2, ],
+  phae_anth_downl <- download_media(metadata = a_zambiana[1:2, ],
   path = tempdir())
 }
 #> ✔ All files were downloaded successfully 🌈
