+++ articles/suwo.md
@@ -613,7 +613,7 @@ head(azam_files, 4)
 fs::dir_tree(path = out_folder)
 ```
 
-     /tmp/RtmpygRQbF/amanita_zambiana 
+     /tmp/RtmpKbrUMy/amanita_zambiana 
     ├──  Amanita_zambiana-GBIF3759537817-1.jpeg 
     ├──  Amanita_zambiana-GBIF3759537817-2.jpeg 
     ├──  Amanita_zambiana-GBIF4430877067-1.jpeg 
@@ -692,7 +692,7 @@ dhol_files <- download_media(metadata = d_holocanthus,
 fs::dir_tree(path = out_folder)
 ```
 
-     /tmp/RtmpygRQbF/diodon_holocanthus 
+     /tmp/RtmpKbrUMy/diodon_holocanthus 
     ├──  Cabo Verde 
     │   └──  Diodon_holocanthus-GBIF3985886532.jpeg 
     ├──  Cayman Islands 
@@ -776,7 +776,7 @@ Click to see
 
     loaded via a namespace (and not attached):
      [1] viridisLite_0.4.2      farver_2.1.2           blob_1.2.4             viridis_0.6.5         
-     [5] S7_0.2.0               fastmap_1.2.0          digest_0.6.37          rpart_4.1.24          
+     [5] S7_0.2.0               fastmap_1.2.0          digest_0.6.38          rpart_4.1.24          
      [9] timechange_0.3.0       lifecycle_1.0.4        survival_3.8-3         RSQLite_2.4.4         
     [13] magrittr_2.0.4         compiler_4.5.2         rlang_1.1.6            sass_0.4.10           
     [17] tools_4.5.2            yaml_2.3.10            data.table_1.17.8      htmlwidgets_1.6.4     
