### 2025-11-13 8e230a5
added key to pkgcheck workflow
