### 2025-11-17 53b614f
addded back badges
