+++ tests/testthat/test_map_locations.R
@@ -1,13 +1,6 @@
 test_that("map Helicobacter pylori", {
-  skip_on_cran()
-  skip_if_offline()
 
-  df1 <- query_inaturalist(species = "Helicobacter pylori", format =  "image",
-                           verbose = TRUE)
-
-  skip_if(is.null(df1))
-
-  df1$latitude[1] <- NA
+  df1 <- suwo:::vignette_metadata$t_rufiventris
 
   a <- map_locations(df1, cluster = TRUE)
 
