+++ NAMESPACE
@@ -26,6 +26,16 @@ importFrom(kableExtra,cell_spec)
 importFrom(kableExtra,column_spec)
 importFrom(kableExtra,kable_styling)
 importFrom(kableExtra,kbl)
+importFrom(leaflet,JS)
+importFrom(leaflet,addAwesomeMarkers)
+importFrom(leaflet,addEasyButton)
+importFrom(leaflet,addMiniMap)
+importFrom(leaflet,addTiles)
+importFrom(leaflet,awesomeIcons)
+importFrom(leaflet,easyButton)
+importFrom(leaflet,easyButtonState)
+importFrom(leaflet,leaflet)
+importFrom(leaflet,markerClusterOptions)
 importFrom(lubridate,parse_date_time)
 importFrom(methods,is)
 importFrom(parallel,detectCores)
