### 2025-11-18 a3fc103
Update CITATION.cff
