+++ CITATION.cff
@@ -247,6 +247,31 @@ references:
   year: '2025'
   doi: 10.32614/CRAN.package.fs
 - type: software
+  title: leaflet
+  abstract: 'leaflet: Create Interactive Web Maps with the JavaScript ''Leaflet''
+    Library'
+  notes: Imports
+  url: https://rstudio.github.io/leaflet/
+  repository: https://CRAN.R-project.org/package=leaflet
+  authors:
+  - family-names: Cheng
+    given-names: Joe
+    email: joe@posit.co
+  - family-names: Schloerke
+    given-names: Barret
+    email: barret@posit.co
+    orcid: https://orcid.org/0000-0001-9986-114X
+  - family-names: Karambelkar
+    given-names: Bhaskar
+  - family-names: Xie
+    given-names: Yihui
+  - family-names: Aden-Buie
+    given-names: Garrick
+    email: garrick@posit.co
+    orcid: https://orcid.org/0000-0002-7111-0077
+  year: '2025'
+  doi: 10.32614/CRAN.package.leaflet
+- type: software
   title: covr
   abstract: 'covr: Test Coverage for Packages'
   notes: Suggests
@@ -270,31 +295,6 @@ references:
     email: hadley@posit.co
   year: '2025'
   doi: 10.32614/CRAN.package.testthat
-- type: software
-  title: leaflet
-  abstract: 'leaflet: Create Interactive Web Maps with the JavaScript ''Leaflet''
-    Library'
-  notes: Suggests
-  url: https://rstudio.github.io/leaflet/
-  repository: https://CRAN.R-project.org/package=leaflet
-  authors:
-  - family-names: Cheng
-    given-names: Joe
-    email: joe@posit.co
-  - family-names: Schloerke
-    given-names: Barret
-    email: barret@posit.co
-    orcid: https://orcid.org/0000-0001-9986-114X
-  - family-names: Karambelkar
-    given-names: Bhaskar
-  - family-names: Xie
-    given-names: Yihui
-  - family-names: Aden-Buie
-    given-names: Garrick
-    email: garrick@posit.co
-    orcid: https://orcid.org/0000-0002-7111-0077
-  year: '2025'
-  doi: 10.32614/CRAN.package.leaflet
 - type: software
   title: knitr
   abstract: 'knitr: A General-Purpose Package for Dynamic Report Generation in R'
