+++ articles/suwo.md
@@ -50,9 +50,10 @@ Here is a description of each step:
 
 1.  Queries regarding a species are submitted through one of the
     available query functions (`query_repo_name()`) that connect to five
-    different online repositories. The output of these queries is a data
-    frame containing metadata associated with the media files (e.g.,
-    species name, date, location, etc, see below).
+    different online repositories (Xeno-Canto, Inaturalist, GBIF,
+    Macaulay Library and Wikiaves). The output of these queries is a
+    data frame containing metadata associated with the media files
+    (e.g., species name, date, location, etc, see below).
 
 2.  If multiple repositories are queried, the resulting metadata data
     frames can be merged into a single data frame using the
@@ -613,7 +614,7 @@ head(azam_files, 4)
 fs::dir_tree(path = out_folder)
 ```
 
-     /tmp/Rtmp6O3gMq/amanita_zambiana 
+     /tmp/RtmpohV36g/amanita_zambiana 
     ├──  Amanita_zambiana-GBIF3759537817-1.jpeg 
     ├──  Amanita_zambiana-GBIF3759537817-2.jpeg 
     ├──  Amanita_zambiana-GBIF4430877067-1.jpeg 
@@ -692,7 +693,7 @@ dhol_files <- download_media(metadata = d_holocanthus,
 fs::dir_tree(path = out_folder)
 ```
 
-     /tmp/Rtmp6O3gMq/diodon_holocanthus 
+     /tmp/RtmpohV36g/diodon_holocanthus 
     ├──  Cabo Verde 
     │   └──  Diodon_holocanthus-GBIF3985886532.jpeg 
     ├──  Cayman Islands 
@@ -776,25 +777,25 @@ Click to see
 
     loaded via a namespace (and not attached):
      [1] viridisLite_0.4.2      farver_2.1.2           blob_1.2.4             viridis_0.6.5         
-     [5] S7_0.2.0               fastmap_1.2.0          digest_0.6.38          rpart_4.1.24          
-     [9] timechange_0.3.0       lifecycle_1.0.4        survival_3.8-3         RSQLite_2.4.4         
-    [13] magrittr_2.0.4         compiler_4.5.2         rlang_1.1.6            sass_0.4.10           
-    [17] tools_4.5.2            yaml_2.3.10            data.table_1.17.8      htmlwidgets_1.6.4     
-    [21] bit_4.6.0              curl_7.0.0             xml2_1.4.1             RColorBrewer_1.1-3    
-    [25] desc_1.4.3             nnet_7.3-20            grid_4.5.2             xtable_1.8-4          
-    [29] e1071_1.7-16           future_1.67.0          ada_2.0-5              ggplot2_4.0.0         
-    [33] globals_0.18.0         scales_1.4.0           MASS_7.3-65            cli_3.6.5             
-    [37] crayon_1.5.3           rmarkdown_2.30         ragg_1.5.0             generics_0.1.4        
-    [41] rstudioapi_0.17.1      RecordLinkage_0.4-12.5 future.apply_1.20.0    httr_1.4.7            
-    [45] DBI_1.2.3              pbapply_1.7-4          cachem_1.1.0           proxy_0.4-27          
-    [49] stringr_1.6.0          splines_4.5.2          parallel_4.5.2         vctrs_0.6.5           
-    [53] Matrix_1.7-4           jsonlite_2.0.0         bit64_4.6.0-1          listenv_0.10.0        
-    [57] systemfonts_1.3.1      jpeg_0.1-11            evd_2.3-7.1            jquerylib_0.1.4       
-    [61] glue_1.8.0             parallelly_1.45.1      pkgdown_2.2.0          codetools_0.2-20      
[…]
