### 2025-11-19 #729 @mpadge
@maRce10 I suspect the "kill" message is due to a timeout. I ran `testthat::test_local()` in our Docker container with your latest version, and it was still running half a day later. When I killed it, it was still running [https://github.com/maRce10/suwo/blob/main/tests/testthat/test_download_media.R](https://github.com/maRce10/suwo/blob/a3fc1032a25723f554f763f2a52bf58b4a1b3a3e/tests/testthat/test_download_media.R#L407). I also note that a lot of your tests have skip conditions like these:
``` r
  skip_on_cran()
  skip_if_offline()
  skip_if(!nzchar(Sys.getenv("XENO_CANTO_API_KEY")),
          "Xeno-Canto API key not set")
```
You should be able to use test mocking facilities to mock API results used in most tests, rather than resort to these kind of conditions, which will effectively only be able to be run by you (and maybe a CI system with your key). We've got lots of detail in our book https://books.ropensci.org/http-testing. I also note that you're using `httr` for your requests. I'd suggest upgrading from that to [`httr2`](https://httr2.r-lib.org), among other reasons because it will make mocking results easier.

## Next steps for you

_Mandatory:_

- Please ensure that you can run `testthat::test_local()` on your package within any clean Docker container (like [ours](https://github.com/ropensci-review-tools/roreviewapi/blob/main/Dockerfile) or the tidyverse one) and get everything to finish within a reasonable amount of time (like the 10-15 minutes or so you're currently getting on most of your GitHub actions runs).

_Optional:_

These should be considered issues that are likely to arise during review, and so please interpret "_optional_" as a reasonably serious recommendation which is likely to greatly aid the review process.

- Upgrade from `httr` to `httr2`.
- Mock test results to extend coverage without use of API key.
