+++ testing/testing-checking suwo functions.Rmd
@@ -490,6 +490,11 @@ lintr::lint_package(linters =  list(lintr::line_length_linter()))
 codemetar::write_codemeta()
 ```
 
+# check with docker
+```{r}
+sudo docker run --rm ghcr.io/ropensci-review-tools/pkgcheck bash -c "git clone https://github.com/maRce10/suwo.git /suwo && cd /suwo && Rscript -e 'devtools::install(); testthat::test_local()'"
+```
+
 # additional checks
 ```{r}
 pkgcheck::pkgcheck(".", use_cache = FALSE)
