+++ NAMESPACE
@@ -20,6 +20,15 @@ importFrom(cli,cli_bullets)
 importFrom(cli,pluralize)
 importFrom(fs,dir_tree)
 importFrom(graphics,rasterImage)
+importFrom(httr2,req_error)
+importFrom(httr2,req_perform)
+importFrom(httr2,req_user_agent)
+importFrom(httr2,request)
+importFrom(httr2,resp_body_json)
+importFrom(httr2,resp_body_string)
+importFrom(httr2,resp_is_error)
+importFrom(httr2,resp_status)
+importFrom(httr2,resp_status_desc)
 importFrom(jpeg,readJPEG)
 importFrom(jsonlite,fromJSON)
 importFrom(kableExtra,cell_spec)
