+++ articles/suwo.md
@@ -614,7 +614,7 @@ head(azam_files, 4)
 fs::dir_tree(path = out_folder)
 ```
 
-     /tmp/RtmpohV36g/amanita_zambiana 
+     /tmp/Rtmpw559Mk/amanita_zambiana 
     ├──  Amanita_zambiana-GBIF3759537817-1.jpeg 
     ├──  Amanita_zambiana-GBIF3759537817-2.jpeg 
     ├──  Amanita_zambiana-GBIF4430877067-1.jpeg 
@@ -693,7 +693,7 @@ dhol_files <- download_media(metadata = d_holocanthus,
 fs::dir_tree(path = out_folder)
 ```
 
-     /tmp/RtmpohV36g/diodon_holocanthus 
+     /tmp/Rtmpw559Mk/diodon_holocanthus 
     ├──  Cabo Verde 
     │   └──  Diodon_holocanthus-GBIF3985886532.jpeg 
     ├──  Cayman Islands 
@@ -777,25 +777,25 @@ Click to see
 
     loaded via a namespace (and not attached):
      [1] viridisLite_0.4.2      farver_2.1.2           blob_1.2.4             viridis_0.6.5         
-     [5] S7_0.2.1               fastmap_1.2.0          leaflet_2.2.3          digest_0.6.38         
+     [5] S7_0.2.1               fastmap_1.2.0          leaflet_2.2.3          digest_0.6.39         
      [9] rpart_4.1.24           timechange_0.3.0       lifecycle_1.0.4        survival_3.8-3        
     [13] RSQLite_2.4.4          magrittr_2.0.4         compiler_4.5.2         rlang_1.1.6           
     [17] sass_0.4.10            tools_4.5.2            yaml_2.3.10            data.table_1.17.8     
-    [21] htmlwidgets_1.6.4      curl_7.0.0             bit_4.6.0              xml2_1.4.1            
+    [21] htmlwidgets_1.6.4      curl_7.0.0             bit_4.6.0              xml2_1.5.0            
     [25] RColorBrewer_1.1-3     desc_1.4.3             nnet_7.3-20            grid_4.5.2            
-    [29] xtable_1.8-4           e1071_1.7-16           future_1.67.0          ada_2.0-5             
+    [29] xtable_1.8-4           e1071_1.7-16           future_1.68.0          ada_2.0-5             
     [33] ggplot2_4.0.1          globals_0.18.0         scales_1.4.0           MASS_7.3-65           
     [37] cli_3.6.5              crayon_1.5.3           rmarkdown_2.30         ragg_1.5.0            
     [41] generics_0.1.4         rstudioapi_0.17.1      RecordLinkage_0.4-12.5 future.apply_1.20.0   
-    [45] httr_1.4.7             pbapply_1.7-4          DBI_1.2.3              cachem_1.1.0          
-    [49] proxy_0.4-27           stringr_1.6.0          splines_4.5.2          parallel_4.5.2        
-    [53] vctrs_0.6.5            Matrix_1.7-4           jsonlite_2.0.0         bit64_4.6.0-1         
-    [57] listenv_0.10.0         systemfonts_1.3.1      jpeg_0.1-11            crosstalk_1.2.2       
-    [61] evd_2.3-7.1            jquerylib_0.1.4        glue_1.8.0             parallelly_1.45.1     
-    [65] pkgdown_2.2.0          codetools_0.2-20       lubridate_1.9.4        stringi_1.8.7         
-    [69] gtable_0.3.6           tibble_3.3.0           pillar_1.11.1          htmltools_0.5.8.1     
+    [45] pbapply_1.7-4          DBI_1.2.3              cachem_1.1.0           proxy_0.4-27          
[…]
