### 2025-11-20 #729 @maRce10
Thank you for your help! I upgraded `httr` to `httr2` and ran a clean docker container (on 3 different computers, 2 OS) that didn't find any error. This is code I used to run the  docker container:

`sudo docker run --rm ghcr.io/ropensci-review-tools/pkgcheck bash -c "git clone https://github.com/maRce10/suwo.git /suwo && cd /suwo && Rscript -e 'devtools::install(); testthat::test_local()'"`
