### 2025-11-21 #729 @mpadge
Great work @maRce10 ! The single failing check there is now pretty clear. Can you please let us know your thoughts about mocking some test results to avoid tests only running with API key? That will like also be a great way to extend test coverage. Other than that, I'll hand over to current Editor-in-Chief @jhollist from here on.
