+++ tests/testthat/test_find_duplicates.R
@@ -1,3 +1,6 @@
+options(verbose = TRUE)
+
+
 test_that("find from gbif ml and xc", {
 
   skip_on_cran()
