+++ tests/testthat/test_merge_metadata.R
@@ -1,3 +1,5 @@
+options(verbose = TRUE)
+
 test_that("merge query_wikiaves and xc", {
 
 
