+++ tests/testthat/test_map_locations.R
@@ -1,3 +1,6 @@
+options(verbose = TRUE)
+
+
 test_that("map Helicobacter pylori", {
 
   df1 <- suwo:::vignette_metadata$t_rufiventris
