+++ tests/testthat/test_query_gbif.R
@@ -1,3 +1,5 @@
+options(verbose = TRUE)
+
 test_that("search Glaucis dohrnii audio", {
   skip_on_cran()
   skip_if_offline()
