+++ tests/testthat/test_download_media.R
@@ -1,32 +1,33 @@
+
+options(verbose = TRUE)
+
 ## Xeno-Canto
 test_that("Xenocanto Phaethornis anthophilus download all.data  = FALSE", {
   skip_on_cran()
   skip_if_offline()
-  skip_if(!nzchar(Sys.getenv("XENO_CANTO_API_KEY")),
-          "Xeno-Canto API key not set")
 
-  xc1 <- query_xenocanto(
+  df1 <- query_gbif(
     species = 'Phaethornis anthophilus',
     all_data = FALSE,
-    api_key = Sys.getenv("XENO_CANTO_API_KEY")
+    format = "image"
   )
 
-  skip_if(is.null(xc1))
+  skip_if(is.null(df1))
 
-  test_keys <- c("532163", "568491")
 
-  sxc1 <- subset(xc1, key %in% test_keys)
+  df1 <- df1[df1$key %in%  c("5063794756", "5077057045"), ]
+  sxc1 <- df1[!duplicated(df1$key), ][1:2, ]
 
   a <- download_media(metadata = sxc1, path = tempdir())
 
-  fls <- list.files(path = tempdir(), pattern = "mp3$")
+  fls <- list.files(path = tempdir(), pattern = "jpeg$")
 
   # remove files
   unlink(file.path(tempdir(), fls))
 
   expected_files <- c(
-    "Phaethornis_anthophilus-XC532163.mp3",
-    "Phaethornis_anthophilus-XC568491.mp3"
+    "Phaethornis_anthophilus-GBIF5063794756.jpeg",
+    "Phaethornis_anthophilus-GBIF5077057045.jpeg"
   )
 
   # test
@@ -40,39 +41,35 @@ test_that("Xenocanto Phaethornis anthophilus download all.data  = FALSE", {
 test_that("Xenocanto Phaethornis anthophilus download folder_by", {
   skip_on_cran()
   skip_if_offline()
-  skip_if(!nzchar(Sys.getenv("XENO_CANTO_API_KEY")),
-          "Xeno-Canto API key not set")
 
-  xc1 <- query_xenocanto(
+  df1 <- query_gbif(
     species = 'Phaethornis anthophilus',
     all_data = FALSE,
-    api_key = Sys.getenv("XENO_CANTO_API_KEY")
+    format = "image"
   )
 
-  skip_if(is.null(xc1))
-  test_keys <- c("532163", "568491")
+  skip_if(is.null(df1))
 
-  sxc1 <- subset(xc1, key %in% test_keys)
 
-  a <- download_media(metadata = sxc1,
-                      path = tempdir(),
-                      folder_by = "date")
+  df1 <- df1[df1$key %in%  c("5063794756", "5077057045"), ]
+  sxc1 <- df1[!duplicated(df1$key), ][1:2, ]
+
+  a <- download_media(metadata = sxc1, path = tempdir(), folder_by = "country")
 
   fls <- list.files(path = tempdir(),
-                    pattern = "mp3$",
+                    pattern = "jpeg$",
                     recursive = TRUE)
 
   # remove filess
-  unlink(file.path(tempdir(), fls))
+  unlink(file.path(tempdir(), a$downloaded_file_name))
 
   expected_files <- c(
-    "2020-03-05/Phaethornis_anthophilus-XC532163.mp3",
-    "2020-06-14/Phaethornis_anthophilus-XC568491.mp3"
+    "Panama/Phaethornis_anthophilus-GBIF5063794756.jpeg",
+    "Colombia/Phaethornis_anthophilus-GBIF5077057045.jpeg"
   )
 
   # test
   expect_true(all(expected_files %in% fls))
-
   expect_true(all(a$downloaded_file_name %in% expected_files))
 
 })
[…]
