### 2025-11-21 #729 @maRce10
thanks!

about API mocking:  the API key is only needed for 1 function. I would rather avoid API mocking as some API accessing functions send multiple requests to the API server in a single function call. It's also better to have the function accessing the actual API in order to detect changes in the responses.

I increased the code coverage to 85% including the API key as a secret, but should be around 76% without the key . The docker container found no errors and run for about 8 minutes. Both RCMD check and pkgcheck workflows passed successfully.
