### 2025-11-22 f16b8bc
check internet before downloading and accessing repos
