### 2025-11-25 a7c8edf
fix local file reading in query_macaulay
