+++ R/internal_functions.R
@@ -147,7 +147,8 @@ pblapply_sw_int <- function(X,
                   paste0("Obtaining metadata ({n} matching record{?s} found)"),
                   as = c("success", "warning", "failure", "error", "message"),
                      n = NULL,
-                     suffix = ":\n") {
+                     suffix = ":\n",
+                  nfiles = NULL) {
   if (!is.null(n))
     text <- cli::pluralize(text)
 
