+++ .github/workflows/readme.yml
@@ -20,6 +20,12 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
+
+     - name: Setup Git config
+        run: |
+          git config --global user.name "github-actions[bot]"
+          git config --global user.email "github-actions[bot]@users.noreply.github.com"
+
       - name: Prepare repository
         run: git checkout "${GITHUB_REF:11}"
       - uses: r-lib/actions/setup-pandoc@v2
