### 2025-11-26 6eafc48
Deploying to gh-pages from @ maRce10/suwo@247564595ea7dc1cb185afe22bec5e3cfd000fc0 🚀
