### 2025-11-26 369c34d
update vignette  authors
