+++ R/suwo-package.R
@@ -47,7 +47,7 @@
 #' @importFrom methods is
 #' @importFrom cli cli_alert_info cli_alert_danger cli_alert_warning
 #' cli_alert_success pluralize cli_bullets
-#' @author Marcelo Araya-Salas, Jorge Elizondo-Calvo & Alejandro Rico
+#' @author Marcelo Araya-Salas, Jorge Elizondo-Calvo & Alejandro Rico-Guevara
 #'
 #'   Maintainer: Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
 #'
