+++ README.Rmd
@@ -86,7 +86,7 @@ Take a look at the [package vignette](https://marce10.github.io/suwo/articles/su
 
 ## Intended use and responsible practices
 
-The [suwo](https://marce10.github.io/suwo/) package is designed exclusively for non-commercial, scientific purposes, including research, education, and conservation. **Any commercial use of the data or media retrieved through this package is strictly prohibited** unless explicit, separate permission is granted directly from the original source platforms and content creators. Users are obligated to comply with the specific terms of service and data use policies of each source database, which often mandate attribution and similarly restrict commercial application. The package developers assume no liability for misuse of the retrieved data or violations of third-party terms of service.
+The [suwo](https://marce10.github.io/suwo/) package is designed exclusively for non-commercial, scientific purposes, including research, education, and conservation. **Any commercial use of the data or media retrieved through this package is strictly prohibited** unless explicit, separate permission is granted directly from the original source platforms and rights holders. Users are obligated to comply with the specific terms of service and data use policies of each source database, which often mandate attribution and similarly restrict commercial application. The package developers assume no liability for misuse of the retrieved data or violations of third-party terms of service.
 
 ## Citation
 
