+++ vignettes/suwo.Rmd
@@ -3,7 +3,7 @@ title: package overview
 subtitle: "suwo: access nature media repositories"
 pagetitle: suwo package overview
 author: 
-- Jorge Elizondo, <a href="https://marce10.github.io/">Marcelo Araya-Salas</a> & <a href="https://marce10.github.io/">Marcelo Araya-Salas</a>
+- Jorge Elizondo, <a href="https://marce10.github.io/">Marcelo Araya-Salas</a> & <a href="https://ecophysics.org/">Alejandro Rico-Guevara</a>
 date:  "`r Sys.Date()`"
 output:
   rmarkdown::html_document:
@@ -103,7 +103,7 @@ for non-commercial, scientific purposes, including research, education, and
 conservation. Any 
 **commercial use of the data or media retrieved through this package is strictly prohibited** 
 unless explicit, separate permission is granted directly from the original 
-source platforms and content creators. Users are obligated to comply with the 
+source platforms and rights holders. Users are obligated to comply with the 
 specific terms of service and data use policies of each source database, which 
 often mandate attribution and similarly restrict commercial application. 
 The package developers assume no liability for misuse of the retrieved data 
@@ -447,7 +447,7 @@ Users must click on the "Export" button to save the .csv file with the metadata:
 
 Here are some tips for using this function properly:
 
-* Valid species names can be checked at `suwo:::ml_taxon_code$SCI_NAME`
+* Valid bird species names can be checked at `suwo:::ml_taxon_code$SCI_NAME` (see 'Querying non-bird species' section below for other groups)
 * The exported csv file must be saved in the directory specified by the `path` argument of the function (default is the current working directory)
 * If the file is saved overwriting a pre-existing file (i.e. same file name) the function will not detect it
 * The function will not proceed until the file is saved 
@@ -484,9 +484,9 @@ Users can check at the Macaulay Library website how many records are available f
 
 <center><img src="ml_num_recs.jpeg" alt="Checking number of records at Macaulay Library" width="100%"></center>
 
-### Querying non-bird species
+### Querying non-bird species 
 
-[query_macaulay()](https://marce10.github.io/suwo/reference/query_macaulay.html) can only search for bird species. However, the Macaulay Library hosts media files for other taxonomic groups as well (e.g., mammals, amphibians, insects). To obtain metadata for these other groups (e.g., mammals, amphibians, insects) users must manually search for the species at the [Macaulay Library website](https://www.macaulaylibrary.org/), download the .csv file with the metadata, and then provide the name of the csv file(s) to the argument `files`.
+[query_macaulay()](https://marce10.github.io/suwo/reference/query_macaulay.html) can only search for bird species. However, the Macaulay Library hosts media files for other taxonomic groups as well (e.g., mammals, amphibians, insects). To obtain metadata for these other groups (e.g., mammals, amphibians, insects) users must manually search for the species at the [Macaulay Library website](https://www.macaulaylibrary.org/), download the .csv file with the metadata, and then provide 1) the name of the csv file(s) to the argument `files` and 2) the directory path were it was saved to the argument `path`.
 
 ## query_xenocanto()
 
