### 2025-11-27 becccf5
fix suwo:::query error when suwo not loaded
