+++ tests/testthat/test_query_xenocanto.R
@@ -130,7 +130,7 @@ test_that("test raw_data TRUE", {
 
   skip_if(is.null(df1))
 
-  expect_true(ncol(df1) == 45)
+  expect_true(ncol(df1) >= 40)
 })
 
 
