+++ vignettes/suwo.Rmd
@@ -448,7 +448,7 @@ Users must click on the "Export" button to save the .csv file with the metadata:
 Here are some tips for using this function properly:
 
 * Valid bird species names can be checked at `suwo:::ml_taxon_code$SCI_NAME` (see 'Querying non-bird species' section below for other groups)
-* The exported csv file must be saved in the directory specified by the `path` argument of the function (default is the current working directory)
+* The exported csv file must be saved in the directory specified by the argument `path` of the function (default is the current working directory)
 * If the file is saved overwriting a pre-existing file (i.e. same file name) the function will not detect it
 * The function will not proceed until the file is saved 
 * Users must log in to the Macaulay Library/eBird account in order to access large batches of observations
@@ -468,7 +468,7 @@ subset(vignette_metadata$p_striigularis,
 
 ### Bypassing record limit
 
-Even if logged in, a maximum of 10000 records per query can be returned. This can be bypassed by using the 'dates' argument to split the search into a sequence of shorter date ranges. The rationale is that by splitting the search into date ranges, users can download multiple .csv files, which are then combined by the function into a single metadata data frame. Of course users must download the csv for each data range.  The following code looks for photos of costa's hummingbird (_Calypte costae_). As Macaulay Library hosts more than 30000 costa's hummingbird records, we need to split the query into multiple date ranges:
+Even if logged in, a maximum of 10000 records per query can be returned. This can be bypassed by using the argument `dates` to split the search into a sequence of shorter date ranges. The rationale is that by splitting the search into date ranges, users can download multiple .csv files, which are then combined by the function into a single metadata data frame. Of course users must download the csv for each data range.  The following code looks for photos of costa's hummingbird (_Calypte costae_). As Macaulay Library hosts more than 30000 costa's hummingbird records, we need to split the query into multiple date ranges:
 
 ```{r, eval = FALSE}
 # test a query with more than 10000 results paging by date
@@ -632,7 +632,7 @@ Note that the [find_duplicates()](https://marce10.github.io/suwo/reference/find_
 subset(dups_merged_metadata, duplicate_group == boesman_group)
 ```
 
-Also note that the locality is not exactly the same for these records, but the fuzzy matching approach used by [find_duplicates()](https://marce10.github.io/suwo/reference/find_duplicates.html) was able to identify them as potential duplicates. By default, the criteria is set to `country > 0.8 & locality > 0.5 & user_name > 0.8 & time == 1 & date == 1` which means that two entries will be considered duplicates if they have a country similarity greater than 0.8, locality similarity greater than 0.5, user_name similarity greater than 0.8, and exact matches for time and date (similarities range from 0 to 1). These values have been found to work well in most cases. Nonetheless, users can adjust the sensitivity based on their specific needs using the argument 'criteria'.
+Also note that the locality is not exactly the same for these records, but the fuzzy matching approach used by [find_duplicates()](https://marce10.github.io/suwo/reference/find_duplicates.html) was able to identify them as potential duplicates. By default, the criteria is set to `country > 0.8 & locality > 0.5 & user_name > 0.8 & time == 1 & date == 1` which means that two entries will be considered duplicates if they have a country similarity greater than 0.8, locality similarity greater than 0.5, user_name similarity greater than 0.8, and exact matches for time and date (similarities range from 0 to 1). These values have been found to work well in most cases. Nonetheless, users can adjust the sensitivity based on their specific needs using the argument `criteria`.
 
[…]
