### 2025-11-28 575cec0
added 'taxon_code' argument to query_macaulay() and fix file monitoring issues
