+++ R/query_inaturalist.R
@@ -108,7 +108,7 @@ query_inaturalist <- function(species = getOption("species"),
     cl <- cores
   }
 
-  query_output_list <- pblapply_sw_int(offsets, cl = cl, pbar = pb,
+  query_output_list <- .pbapply_sw(offsets, cl = cl, pbar = pb,
                                        function(offset) {
     query_output <-
       try(jsonlite::fromJSON(paste0(base_url, "&offset=", offset)),
