+++ R/download_media.R
@@ -123,7 +123,7 @@ download_media <-
     }
 
     metadata$download_status <-
-      unlist(pblapply_sw_int(
+      unlist(.pbapply_sw(
         pbar = pb,
         X = seq_len(nrow(metadata)),
         cl = cl,

[…]
