+++ man/query_macaulay.Rd
@@ -6,6 +6,7 @@
 \usage{
 query_macaulay(
   species = getOption("species"),
+  taxon_code = NULL,
   format = c("sound", "image", "video"),
   verbose = getOption("verbose", TRUE),
   all_data = getOption("all_data", FALSE),
@@ -21,6 +22,9 @@ query_macaulay(
 the format: "Genus epithet". Required. Can be set globally for the current
 R session via the "term" option
 (e.g. \code{options(term = "Hypsiboas rufitelus")}).}
+
+\item{taxon_code}{Optional character string with the Macaulay Library taxon
+code (see vignette for more details). If provided, 'species' is ignored.}
 
 \item{format}{Character vector with the media format to query for. Options
 are 'sound', 'image' of 'video'. Required.}
@@ -49,7 +53,8 @@ is saved into the current working directory (\code{"."}).}
 
 \item{files}{Optional character vector with the name(s) of the .csv file(s)
 to read. If provided, the function will import the data from the .csv files
-instead of opening the Macaulay Library search page in a browser.}
+instead of opening the Macaulay Library search page in a browser ('species'
+is ignored if supplied).}
 
 \item{dates}{Optional numeric vector with years to split the search. If
 provided, the function will perform separate queries for each date range
