+++ man/suwo-package.Rd
@@ -45,7 +45,7 @@ Useful links:
 
 }
 \author{
-Marcelo Araya-Salas, Jorge Elizondo-Calvo & Alejandro Rico
+Marcelo Araya-Salas, Jorge Elizondo-Calvo & Alejandro Rico-Guevara
 
   Maintainer: Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
 }
