+++ testing/testing-checking suwo functions.Rmd
@@ -449,6 +449,7 @@ sim_annotation <- function(n = 10, n.sound.files = 1){
 ```{r check, eval = FALSE}
 
 #run document twice
+if (!nzchar(Sys.getenv("XENO_CANTO_API_KEY")))
 Sys.setenv(XENO_CANTO_API_KEY = getPass::getPass())
 devtools::document(".")
 
@@ -470,6 +471,7 @@ devtools::check(document = TRUE, run_dont_test = TRUE, vignettes = TRUE, manual 
 # devtools::install()
 
 ## set XC Key
+if (!nzchar(Sys.getenv("XENO_CANTO_API_KEY")))
 Sys.setenv(XENO_CANTO_API_KEY = getPass::getPass())
 
   run.all()
