### 2025-12-05 0b7906c
fix tests .date_ranges
