+++ tests/testthat/test_internals.R
@@ -8,7 +8,7 @@ test_that(".date_ranges", {
 
   dts <- suwo:::.date_ranges(x = seq(1976, 2026, length.out = 50))
 
-  expect_true(nrow(dts) == 94)
+  expect_true(nrow(dts) == 95)
 
 })
 
