+++ vignettes/suwo.Rmd
@@ -150,10 +150,14 @@ involved:
 </br>
 Here is a description of each step:
 
+Obtain metadata: 
+
 1. Queries regarding a species are submitted through one of the available 
 query functions (`query_repo_name()`) that connect to five different online 
 repositories (Xeno-Canto, Inaturalist, GBIF, Macaulay Library and Wikiaves). The output of these queries is a data frame containing metadata 
 associated with the media files (e.g., species name, date, location, etc, see below). 
+
+Curate metadata: 
 
 1. If multiple repositories are queried, the resulting metadata data frames can 
 be merged into a single data frame using the 
