+++ articles/suwo.md
@@ -48,6 +48,8 @@ workflow and the main functions involved:
 
 Here is a description of each step:
 
+Obtain metadata:
+
 1.  Queries regarding a species are submitted through one of the
     available query functions (`query_repo_name()`) that connect to five
     different online repositories (Xeno-Canto, Inaturalist, GBIF,
@@ -55,12 +57,14 @@ Here is a description of each step:
     data frame containing metadata associated with the media files
     (e.g., species name, date, location, etc, see below).
 
-2.  If multiple repositories are queried, the resulting metadata data
+Curate metadata:
+
+1.  If multiple repositories are queried, the resulting metadata data
     frames can be merged into a single data frame using the
     [merge_metadata()](https://marce10.github.io/suwo/reference/merge_metadata.html)
     function.
 
-3.  Check for duplicate records in their datasets using the
+2.  Check for duplicate records in their datasets using the
     [find_duplicates()](https://marce10.github.io/suwo/reference/find_duplicates.html)
     function. Candidate duplicated entries are identified based on
     matching species name, country, date, user name, and geographic
@@ -68,11 +72,11 @@ Here is a description of each step:
     decide which records to keep, which can be done with
     [remove_duplicates()](https://marce10.github.io/suwo/reference/remove_duplicates.html).
 
-4.  Download the media files associated with the metadata using the
+3.  Download the media files associated with the metadata using the
     [download_media()](https://marce10.github.io/suwo/reference/download_media.html)
     function.
 
-5.  Users can update their datasets with new records using the
+4.  Users can update their datasets with new records using the
     [update_metadata()](https://marce10.github.io/suwo/reference/update_metadata.html)
     function.
 
@@ -633,7 +637,7 @@ head(azam_files, 4)
 fs::dir_tree(path = out_folder)
 ```
 
-     /tmp/RtmpID7djE/amanita_zambiana 
+     /tmp/RtmpRt5VMN/amanita_zambiana 
     ├──  Amanita_zambiana-GBIF3759537817-1.jpeg 
     ├──  Amanita_zambiana-GBIF3759537817-2.jpeg 
     ├──  Amanita_zambiana-GBIF4430877067-1.jpeg 
@@ -712,7 +716,7 @@ dhol_files <- download_media(metadata = d_holocanthus,
 fs::dir_tree(path = out_folder)
 ```
 
-     /tmp/RtmpID7djE/diodon_holocanthus 
+     /tmp/RtmpRt5VMN/diodon_holocanthus 
     ├──  Cabo Verde 
     │   └──  Diodon_holocanthus-GBIF3985886532.jpeg 
     ├──  Cayman Islands 
@@ -798,9 +802,9 @@ Click to see
      [1] viridisLite_0.4.2      farver_2.1.2           blob_1.2.4             viridis_0.6.5         
      [5] S7_0.2.1               fastmap_1.2.0          leaflet_2.2.3          digest_0.6.39         
      [9] rpart_4.1.24           timechange_0.3.0       lifecycle_1.0.4        survival_3.8-3        
-    [13] RSQLite_2.4.4          magrittr_2.0.4         compiler_4.5.2         rlang_1.1.6           
-    [17] sass_0.4.10            tools_4.5.2            yaml_2.3.10            data.table_1.17.8     
-    [21] htmlwidgets_1.6.4      curl_7.0.0             bit_4.6.0              xml2_1.5.0            
+    [13] RSQLite_2.4.5          magrittr_2.0.4         compiler_4.5.2         rlang_1.1.6           
[…]
