+++ reference/query_macaulay.md
@@ -104,10 +104,7 @@ This is an interactive function which opens a browser window to the
 Macaulay Library's search page, where the user must download a .csv file
 with the metadata. The function then reads the .csv file and returns a
 data frame with the metadata. The function can also import previously
-downloaded metadata (in csv format) with the argument \`files\`. If
-`all_data = TRUE`, all metadata fields (columns) are returned. If
-`raw_data = TRUE`, the raw data as obtained from the repository is
-returned (without any formatting).
+downloaded metadata (in csv format) with the argument \`files\`.
 
 ## Details
 
@@ -115,18 +112,30 @@ This function queries for species observation info in the [Macaulay
 Library](https://www.macaulaylibrary.org/) online repository and returns
 the metadata of media files matching the query. The Macaulay Library is
 the world’s largest repository of digital media (audio, photo, and
-video) of wildlife, and their habitats. The archive hosts more than 77
+video) of wildlife (mostly birds but also other vertebrates and
+invertebrates), and their habitats. The archive hosts more than 77
 million images, 3 million sound recordings, and 350k videos, from more
 than 80k contributors, and is integrated with eBird, the world’s largest
-biodiversity dataset.
+biodiversity dataset. For bird species the species name must be valid
+according to the Macaulay Library taxonomy (which follows the Clements
+checklist). For non-bird species users must use the argument
+\`taxon_code\`. The species taxon code can be found by running a search
+at the [Macaulay Library's search
+page](https://www.macaulaylibrary.org/) and checking the URL of the
+species page. For instance, the URL when searching for jaguar (Panthera
+onca) is
+'https://search.macaulaylibrary.org/catalog?taxonCode=t-11032765' so the
+taxon code is "t-11032765". If `all_data = TRUE`, all metadata fields
+(columns) are returned. If `raw_data = TRUE`, the raw data as obtained
+from the repository is returned (without any formatting). Here are some
+instructions for using this function properly:
 
-Here are some instructions for using this function properly:
-
-- Valid species names can be checked at `suwo:::ml_taxon_code$SCI_NAME`.
+- Valid bird species names can be checked at
+  `suwo:::ml_taxon_code$SCI_NAME`.
 
 - Users must save the save the .csv file manually
 
-- *If the file is saved overwritting a pre-existing file (i.e. same file
+- *If the file is saved overwriting a pre-existing file (i.e. same file
   name) the function will not detect it*
 
 - A maximum of 10000 records per query can be returned, but this can be
@@ -185,5 +194,10 @@ new_clements <- read.csv(clements_url)
 # provide "updated" clements list to query_macaulay()
 tur_ili2 <- query_macaulay(species = "Turdus iliacus", format = "sound",
  taxon_code_info = new_clements, path = tempdir())
+
+ # query using taxon code
+ # this is the URL when querying jaguars:
+ # https://search.macaulaylibrary.org/catalog?taxonCode=t-11032765
+ p_onca <- query_macaulay(taxon_code = "t-11032765", format = "image")
 }
 ```
