### 2025-12-06 #729 @adamhsparks
Also, I tried redocumenting, thinking maybe it was an issue with an older version of roxygen being used than what I have installed. When I redocument to update the .Rd files, I get this error as well.

```r
✖ In topic 'template_params.Rd': Skipping; no name and/or title.
```
