+++ R/template_params.R
@@ -30,4 +30,5 @@
 #' @param metadata data frame previously obtained from any suwo query
 #' function (i.e. `query_reponame()`).
 #' @name template_params
+#' @title template_params
 NULL
