### 2025-12-08 c6819ab
fixing pkgdown error template_params
