+++ reference/template_params.md
@@ -1,0 +1,57 @@
+# template_params
+
+template_params
+
+## Arguments
+
+- cores:
+
+  Numeric vector of length 1. Controls whether parallel computing is
+  applied by specifying the number of cores to be used. Default is 1
+  (i.e. no parallel computing). Can be set globally for the current R
+  session via the "mc.cores" option (e.g. `options(mc.cores = 2)`). Note
+  that some repositories might not support parallel queries from the
+  same IP address as it might be identified as denial-of-service
+  cyberattack.
+
+- pb:
+
+  Logical argument to control if progress bar is shown. Default is
+  `TRUE`. Can be set globally for the current R session via the "pb"
+  option ( `options(pb = TRUE)`).
+
+- species:
+
+  Character string with the scientific name of a species in the format:
+  "Genus epithet". Required. Can be set globally for the current R
+  session via the "term" option (e.g.
+  `options(term = "Hypsiboas rufitelus")`).
+
+- verbose:
+
+  Logical argument that determines if text is shown in console. Default
+  is `TRUE`. Can be set globally for the current R session via the
+  "verbose" option ( `options(verbose = TRUE)`).
+
+- all_data:
+
+  Logical argument that determines if all data available from database
+  is shown in the results of search. Default is `FALSE`. Can be set
+  globally for the current R session via the "all_data" option (
+  `options(all_data = TRUE)`).
+
+- raw_data:
+
+  Logical argument that determines if the raw data from the API is
+  returned (e.g. without any manipulation). Default is `FALSE`. Can be
+  set globally for the current R session via the "raw_data" option (
+  `options(raw_data = TRUE)`). If `TRUE` `all_data` is set to `TRUE`
+  internally. Useful for developers, or if users suspect that some data
+  is mishandled during processing (i.e. date information is lost). Note
+  that the metadata obtained when `raw_data = TRUE` is not standardized,
+  so most suwo functions for downstream steps will not work on them.
+
+- metadata:
+
+  data frame previously obtained from any suwo query function (i.e.
+  \`query_reponame()\`).
