### 2025-12-09 #729 @adamhsparks
I've found it. You have extra "{}" in/around your `#' @references` in these six files.

Please remove them and the error message goes away as they are not necessary.
