### 2025-12-09 09fa24c
remove braces in @reference
