+++ R/query_wikiaves.R
@@ -24,11 +24,11 @@
 #' p_nattereri <- query_wikiaves(species = "Phaethornis nattereri",
 #'     format = "image")
 #'
-#' @references {
+#' @references
 #' Schubert, Stephanie Caroline, Lilian Tonelli Manica, and André De Camargo
 #' Guaraldo. 2019. Revealing the potential of a huge citizen-science platform
 #' to study bird migration. Emu-Austral Ornithology 119.4: 364-373.
-#' }
+#'
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
 
 query_wikiaves <-
