+++ R/query_inaturalist.R
@@ -28,9 +28,9 @@
 #' # search Bleeding Tooth mushroom images
 # h_peckii <- query_inaturalist(species = 'Hydnellum peckii', format = "image")
 #'
-#' @references {
+#' @references
 #' iNaturalist. Available from https://www.inaturalist.org. Accessed [date]
-#' }
+#'
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
 #'
 
