+++ R/map_locations.R
@@ -25,11 +25,11 @@
 #' map_locations(e_hochs)
 #' }
 #'
-#' @references {
+#' @references
 #' Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to
 #' streamline analysis of animal acoustic signals. Methods in Ecology and
 #' Evolution, 8(2), 184-191.
-#' }
+#'
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr}) and Grace
 #' Smith Vidaurre
 
