+++ man/query_xenocanto.Rd
@@ -104,10 +104,8 @@ api_key = XC_API_KEY)
 
 }
 \references{
-{
 Planqué, Bob, & Willem-Pier Vellinga. 2008. Xeno-canto: a 21st-century way
 to appreciate Neotropical bird song. Neotrop. Birding 3: 17-23.
-}
 }
 \seealso{
 \code{\link{query_gbif}}, \code{\link{query_wikiaves}},
