+++ man/query_inaturalist.Rd
@@ -83,9 +83,7 @@ field.
 
 }
 \references{
-{
 iNaturalist. Available from https://www.inaturalist.org. Accessed [date]
-}
 }
 \author{
 Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
