### 2025-12-09 #729 @maRce10
Ok I removed the extra braces. Would be useful to know how come I cannot spot them.
