+++ R/query_inaturalist.R
@@ -43,6 +43,7 @@ query_inaturalist <- function(species = getOption("species"),
                               verifiable = FALSE,
                               all_data = getOption("all_data", FALSE),
                               raw_data = getOption("raw_data", FALSE)) {
+
   arguments <- as.list(base::match.call())[-1]
 
   for (i in names(arguments)) {
@@ -55,38 +56,29 @@ query_inaturalist <- function(species = getOption("species"),
   format <- rlang::arg_match(format)
   inat_format <- switch(format, sound = "sounds", image = "photos")
 
-  # Use the unified connection checker
+  # Connection check
   if (!.checkconnection(verb = verbose, service = "inat")) {
     return(invisible(NULL))
   }
 
-
   base_url <- paste0(
     "https://api.inaturalist.org/v1/observations?per_page=200&",
-    "taxon_name=",
-    gsub(" ", "%20", species),
-    "&",
-    inat_format,
-    "=true",
-    "&",
-    "identified=",
-    identified,
-    "&",
-    "verifiable=",
-    verifiable
+    "taxon_name=", gsub(" ", "%20", species), "&",
+    inat_format, "=true&",
+    "identified=", identified, "&",
+    "verifiable=", verifiable, "&",
+    "order=asc&order_by=id"
   )
-
 
   first_query <- try(jsonlite::fromJSON(base_url), silent = TRUE)
 
   # let user know error when downloading metadata
   if (.is_error(first_query)) {
     if (verbose) {
-      .message(text = "Metadata could not be dowloaded", as = "failure")
+      .message(text = "Metadata could not be downloaded", as = "failure")
     }
     return(invisible(NULL))
   }
-
 
   total_results <- first_query$total_results
 
@@ -101,7 +93,7 @@ query_inaturalist <- function(species = getOption("species"),
     .message(n = total_results, as = "success")
   }
 
-  offsets <- seq(0, total_results, by = 200)
+  offsets <- seq(0, max(0, total_results - 1), by = 200)
 
   if (Sys.info()[1] == "Windows" && cores > 1) {
     cl <- parallel::makePSOCKcluster(getOption("cl.cores", cores))
@@ -109,47 +101,56 @@ query_inaturalist <- function(species = getOption("species"),
     cl <- cores
   }
 
-  query_output_list <- .pbapply_sw(offsets, cl = cl, pbar = pb,
-                                       function(offset) {
-    query_output <-
-      try(jsonlite::fromJSON(paste0(base_url, "&offset=", offset)),
-          silent = TRUE)
+  query_output_list <- .pbapply_sw(
+    offsets,
+    cl = cl,
+    pbar = pb,
+    function(offset) {
 
-    if (is.null(query_output$results)) {
-      return(NULL)
+      query_output <- try(
+        jsonlite::fromJSON(paste0(base_url, "&offset=", offset)),
+        silent = TRUE
+      )
+
+      if (is.null(query_output$results)) {
+        return(NULL)
+      }
+
+      if (.is_error(query_output)) {
+        return(query_output)
+      }
+
+      query_output$results <- lapply(
+        seq_len(nrow(query_output$results)),
+        function(u) {
+
+          x <- as.data.frame(query_output$results[u, ])
+
+          media_df <- if (format == "sound")
+            do.call(rbind, x$sounds)
+          else
+            do.call(rbind, x$photos)
+
+          media_df <- media_df[!vapply(media_df, is.list, logical(1))]
+          media_df <- data.frame(media_df)
+
+          x <- x[!vapply(x, is.list, logical(1))]
+          X_df <- data.frame(t(unlist(x)))
+          X_df <- cbind(X_df, media_df)
+
+          return(X_df)
+        }
+      )
+
[…]
