+++ articles/suwo.md
@@ -116,12 +116,12 @@ h_sarapiquensis <- query_inaturalist(species = "Heliconia sarapiquensis",
 head(h_sarapiquensis, 4)
 ```
 
-| repository  | format |    key    |         species         |    date    | time  |    user_name     | country |           locality           | latitude | longitude |                                    file_url                                     | file_extension |
-|:-----------:|:------:|:---------:|:-----------------------:|:----------:|:-----:|:----------------:|:-------:|:----------------------------:|:--------:|:---------:|:-------------------------------------------------------------------------------:|:--------------:|
-| iNaturalist | image  | 263417773 | Heliconia sarapiquensis | 2025-02-28 | 14:23 | Original Madness |   NA    | 10.163116739,-83.9389050007  | 10.16312 | -83.93891 | <https://inaturalist-open-data.s3.amazonaws.com/photos/473219810/original.jpeg> |      jpeg      |
-| iNaturalist | image  | 263417128 | Heliconia sarapiquensis | 2025-02-28 | 14:18 | Original Madness |   NA    | 10.163116739,-83.9389050007  | 10.16312 | -83.93891 | <https://inaturalist-open-data.s3.amazonaws.com/photos/473218501/original.jpeg> |      jpeg      |
-| iNaturalist | image  | 263415801 | Heliconia sarapiquensis | 2025-02-28 | 14:11 | Original Madness |   NA    | 10.163116739,-83.9389050007  | 10.16312 | -83.93891 | <https://inaturalist-open-data.s3.amazonaws.com/photos/473216124/original.jpeg> |      jpeg      |
-| iNaturalist | image  | 234065037 | Heliconia sarapiquensis | 2024-08-05 | 16:57 | Shakiraprovasoli |   NA    | 10.4436533293,-84.0696545598 | 10.44365 | -84.06965 |         <https://static.inaturalist.org/photos/416318291/original.jpeg>         |      jpeg      |
+| repository  | format |    key    |         species         |    date    | time  |         user_name          | country |           locality           | latitude | longitude |                                    file_url                                     | file_extension |
+|:-----------:|:------:|:---------:|:-----------------------:|:----------:|:-----:|:--------------------------:|:-------:|:----------------------------:|:--------:|:---------:|:-------------------------------------------------------------------------------:|:--------------:|
+| iNaturalist | image  | 17899854  | Heliconia sarapiquensis | 2005-09-22 | 19:40 | Gonzalo Mucientes Sandoval |   NA    |     10.46196,-84.737482      | 10.46196 | -84.73748 |  <https://inaturalist-open-data.s3.amazonaws.com/photos/27338069/original.jpg>  |      jpeg      |
+| iNaturalist | image  | 24247964  | Heliconia sarapiquensis | 2019-03-13 | 12:46 |        Lena Struwe         |   NA    | 10.1745970958,-83.9147372247 | 10.17460 | -83.91474 | <https://inaturalist-open-data.s3.amazonaws.com/photos/37493584/original.jpeg>  |      jpeg      |
+| iNaturalist | image  | 29151576  | Heliconia sarapiquensis | 2016-10-16 | 08:58 |        Pete Woodall        |   NA    |       10.1736,-83.9134       | 10.17360 | -83.91340 | <https://inaturalist-open-data.s3.amazonaws.com/photos/45484602/original.jpeg>  |      jpeg      |
[…]
