### 2025-12-17 56bc930
fix inat paging bug and added messages to finding/removing duplicates
