+++ R/find_duplicates.R
@@ -2,6 +2,7 @@
 #'
 #' \code{find_duplicates} detect possible duplicated entries from merged
 #' metadata from several repositories.
+#' @inheritParams template_params
 #' @param metadata data frame obtained from combining the output metadata of
 #' two or more suwo query function using the `merge_metadata()` function.
 #' Data frames obtained from a single suwo query function can also be used
@@ -9,8 +10,9 @@
 #' frame must have the following columns: `user_name`, `locality`, `repository`,
 #'  `country`, `format`, `time`, and `date`.
 #' @param sort Logical argument indicating if the output data frame should be
-#' sorted by the `duplicate_group` column. This will group all potential
-#' duplicates together in the output data frame. Default is `TRUE`.
+#' sorted by the `duplicate_group` column added by this function. This will
+#' group all potential duplicates together in the output data frame.
+#' Default is `TRUE`.
 #' @param criteria A character string indicating the criteria to use to
 #' determine duplicates. By default, the criteria is set to
 #' \code{country > 0.8 & locality > 0.5 & user_name > 0.8 & time == 1
@@ -21,27 +23,28 @@
 #' (similarities range from 0 to 1). These values have been found to work well
 #' in most cases. Users can modify this string to adjust the sensitivity of the
 #' duplicate detection based on their specific needs.
-#' @return A single data frame with the data from all input data frames
-#' combined and with an additional column named `duplicate_group` indicating
-#' potential duplicates with a common index. Entries without potential
-#' duplicates are labeled as `NA` in this new column.
+#' @return A data frame with the input data frame and an additional column
+#' named `duplicate_group` indicating potential duplicates with a common index.
+#' Entries without potential duplicates are labeled as `NA` in this new column.
 #' @export
 #' @name find_duplicates
 #' @details This function compares the information in the entries of a
-#' combined metadata data frame and labels those possible duplicates with
+#' combined metadata data frame (typically the output of
+#' \code{\link{merge_metadata}}) and labels those possible duplicates with
 #' a common index in a new column named `duplicate_group`. The comparison is
 #' based on the similarity of the following fields: `user_name`, `locality`,
 #' `time` and `country`. Only rows with no missing data for those fields will
 #' be considered. The function uses the `RecordLinkage` package to perform the
 #' a fuzzy matching comparison and identify potential duplicates based on
-#' predefined similarity thresholds. The function only spots duplicates from
-#' different repositories and assumes those duplicates should have the same
-#' `format` and `date`. This function is useful for curating the data obtained
-#' by merging data from multiple sources, as the same observation might be
-#' recorded in different repositories. This is a common issue in citizen
-#' science repositories, where users might upload the same observation to
-#' different platforms. This can also occur as some repositories automatically
-#' share data with other repositories, particularly with GBIF.
+#' predefined similarity thresholds (see argument 'criteria'). The function
+#' only spots duplicates from different repositories and assumes those
+#' duplicates should have the same `format` and `date`. This function is
+#' useful for curating the data obtained by merging data from multiple sources,
+#' as the same observation might be recorded in different repositories.
+#' This is a common issue in citizen science repositories, where users might
+#' upload the same observation to different platforms. This can also occur as
+#' some repositories automatically share data with other repositories,
+#' particularly with GBIF.
 #' @examples
 #' # get metadata from 2 repos
 #' gb <- query_gbif(species = "Turdus rufiventris", format =  "sound")
@@ -61,7 +64,8 @@
 #'
 find_duplicates <- function(metadata, sort = TRUE,
    criteria =
[…]
