+++ R/template_params.R
@@ -13,17 +13,18 @@
 #' (e.g. \code{options(term = "Hypsiboas rufitelus")}).
 #' @param verbose Logical argument that determines if text is shown in
 #' console. Default is \code{TRUE}. Can be set globally for the current R
-#' session via the "verbose" option ( \code{options(verbose = TRUE)}).
+#' session via the "verbose" option
+#' ( \code{options(verbose = TRUE)}).
 #' @param all_data Logical argument that determines if all data available
 #' from database is shown in the results of search. Default is \code{FALSE}.
 #' Can be set globally for the current R session via the "all_data"
 #' option ( \code{options(all_data = TRUE)}).
 #' @param raw_data Logical argument that determines if the raw data from the
-#' API is returned (e.g. without any manipulation). Default is \code{FALSE}.
-#' Can be set globally for the current R session via the "raw_data"
-#' option ( \code{options(raw_data = TRUE)}). If \code{TRUE}  \code{all_data}
-#' is set to \code{TRUE} internally. Useful for developers, or if users
-#' suspect that some data is mishandled during processing
+#' repository is returned (e.g. without any manipulation).
+#' Default is \code{FALSE}. Can be set globally for the current R session
+#' via the "raw_data" option ( \code{options(raw_data = TRUE)}). If \code{TRUE}
+#' \code{all_data} is set to \code{TRUE} internally. Useful for developers,
+#' or if users suspect that some data is mishandled during processing
 #' (i.e. date information is lost). Note that the metadata obtained
 #' when \code{raw_data = TRUE} is not standardized, so most suwo functions
 #' for downstream steps will not work on them.
