+++ llms.txt
@@ -93,15 +93,15 @@ nature media repositories. R package version 0.1.0.
 Obtain media metadata from various biodiversity repositories
 
 - [`query_gbif()`](https://marce10.github.io/suwo/reference/query_gbif.md)
-  : Access 'gbif' recordings and metadata
+  : Access 'gbif' media file metadata
 - [`query_inaturalist()`](https://marce10.github.io/suwo/reference/query_inaturalist.md)
-  : Access 'inaturalist' recordings and metadata
+  : Access 'iNaturalist' media file metadata
 - [`query_macaulay()`](https://marce10.github.io/suwo/reference/query_macaulay.md)
   : Searches for media files in the Macaulay Library
 - [`query_wikiaves()`](https://marce10.github.io/suwo/reference/query_wikiaves.md)
-  : Access 'Wikiaves' recordings and metadata
+  : Access 'WikiAves' media file metadata
 - [`query_xenocanto()`](https://marce10.github.io/suwo/reference/query_xenocanto.md)
-  : Access 'Xeno-Canto' recordings and metadata
+  : Access 'Xeno-Canto' recording metadata
 
 ## Obtain media files
 
@@ -121,9 +121,9 @@ Data exploration and manipulation
 - [`find_duplicates()`](https://marce10.github.io/suwo/reference/find_duplicates.md)
   : Find duplicated entries in metadata
 - [`remove_duplicates()`](https://marce10.github.io/suwo/reference/remove_duplicates.md)
-  : Merge metadata data frames
+  : Remove duplicated media records
 - [`map_locations()`](https://marce10.github.io/suwo/reference/map_locations.md)
-  : Maps of instances of observations by species
+  : Maps of media records
 
 # Articles
 

[…]
