+++ tests/testthat/test_query_wikiaves.R
@@ -46,8 +46,9 @@ test_that("test verbose FALSE", {
   skip_on_cran()
   skip_if_offline()
 
-  df1 <- testthat::capture_output(query_wikiaves(species = 'a3', format =  "sound",
-                                       verbose = FALSE, pb = FALSE))
+  df1 <- testthat::capture_output(query_wikiaves(species = 'a3',
+                                                 format =  "sound",
+                                                 verbose = FALSE, pb = FALSE))
 
   skip_if(is.null(df1))
 
