### 2025-12-25 0960821
Deploying to gh-pages from @ maRce10/suwo@654066f5dd5437053e9d9f0e64ac7c05732efda8 🚀
