### 2026-01-27 4266427
remove .onLoad from internal functions
