### 2026-01-27 c60745f
make "image" default format
