### 2026-01-27 32e67ec
update parallel::makePSOCKcluster(cores)
