+++ R/query_wikiaves.R
@@ -170,7 +170,7 @@ query_wikiaves <-
 
     # set clusters for windows OS
     if (Sys.info()[1] == "Windows" && cores > 1) {
-      cl <- parallel::makePSOCKcluster(getOption("cl.cores", cores))
+      cl <- parallel::makePSOCKcluster(cores)
     } else {
       cl <- cores
     }
