+++ R/remove_duplicates.R
@@ -93,7 +93,7 @@ remove_duplicates <-
 
     # set clusters for windows OS
     if (Sys.info()[1] == "Windows" && cores > 1) {
-      cl <- parallel::makePSOCKcluster(getOption("cl.cores", cores))
+      cl <- parallel::makePSOCKcluster(cores)
     } else {
       cl <- cores
     }
