+++ R/download_media.R
@@ -115,7 +115,7 @@ download_media <-
       write(file = "", x = "Downloading media files:")
     }
     if (Sys.info()[1] == "Windows" && cores > 1) {
-      cl <- parallel::makePSOCKcluster(getOption("cl.cores", cores))
+      cl <- parallel::makePSOCKcluster(cores)
     } else {
       cl <- cores
     }
