+++ R/query_xenocanto.R
@@ -124,7 +124,7 @@ query_xenocanto <-
 
     if (Sys.info()[1] == "Windows" && cores > 1) {
       cl <-
-        parallel::makePSOCKcluster(getOption("cl.cores", cores))
+        parallel::makePSOCKcluster(cores)
     } else {
       cl <- cores
     }
