+++ man/query_wikiaves.Rd
@@ -6,7 +6,7 @@
 \usage{
 query_wikiaves(
   species = getOption("species"),
-  format = getOption("format", c("sound", "image")),
+   format = getOption("format", c("image", "sound")),
   cores = getOption("mc.cores", 1),
   pb = getOption("pb", TRUE),
   verbose = getOption("verbose", TRUE),
