+++ .github/workflows/readme.yml
@@ -3,12 +3,9 @@ on:
   push:
     branches:
     - main
-    - master
     paths:
       - README.Rmd
       - .github/workflows/readme.yml
-  schedule:
-    - cron:  '0 12 * * *'
 
 name: Render README
 
