+++ articles/suwo.md
@@ -612,6 +612,8 @@ dir.create(out_folder)
 azam_files <- download_media(metadata = a_zam, path = out_folder)
 ```
 
+     [32m✔ [39m All files were downloaded successfully 🥇
+
 The output of the function is a data frame similar to the input metadata
 but with two additional columns indicating the file name of the
 downloaded files (‘downloaded_file_name’) and the result of the download
@@ -637,12 +639,18 @@ head(azam_files, 4)
 fs::dir_tree(path = out_folder)
 ```
 
-     /tmp/Rtmp3gaEWv/amanita_zambiana 
+     /tmp/RtmppRoexH/amanita_zambiana 
     ├──  Amanita_zambiana-GBIF3759537817-1.jpeg 
     ├──  Amanita_zambiana-GBIF3759537817-2.jpeg 
     ├──  Amanita_zambiana-GBIF4430877067-1.jpeg 
     ├──  Amanita_zambiana-GBIF4430877067-2.jpeg 
     ├──  Amanita_zambiana-GBIF4430877067-3.jpeg 
+    ├──  Amanita_zambiana-GBIF5069132689-1.jpeg 
+    ├──  Amanita_zambiana-GBIF5069132689-2.jpeg 
+    ├──  Amanita_zambiana-GBIF5069132691.jpeg 
+    ├──  Amanita_zambiana-GBIF5069132696-1.jpeg 
+    ├──  Amanita_zambiana-GBIF5069132696-2.jpeg 
+    ├──  Amanita_zambiana-GBIF5069132732.jpeg 
     └──  Amanita_zambiana-GBIF5104283819.jpeg 
 
 Note that the name of the downloaded files includes the species name, an
@@ -710,13 +718,13 @@ dhol_files <- download_media(metadata = d_holocanthus,
                              folder_by = "country")
 ```
 
-     [32m✔ [39m All files were downloaded successfully 🥇
+     [32m✔ [39m All files were downloaded successfully 🎊
 
 ``` r
 fs::dir_tree(path = out_folder)
 ```
 
-     /tmp/Rtmp3gaEWv/diodon_holocanthus 
+     /tmp/RtmppRoexH/diodon_holocanthus 
     ├──  Cabo Verde 
     │   └──  Diodon_holocanthus-GBIF3985886532.jpeg 
     ├──  Cayman Islands 
@@ -799,26 +807,26 @@ Click to see
     [1] kableExtra_1.4.0 suwo_0.1.0       knitr_1.51      
 
     loaded via a namespace (and not attached):
-     [1] viridisLite_0.4.2      farver_2.1.2           blob_1.2.4             viridis_0.6.5         
+     [1] viridisLite_0.4.2      farver_2.1.2           blob_1.3.0             viridis_0.6.5         
      [5] S7_0.2.1               fastmap_1.2.0          leaflet_2.2.3          digest_0.6.39         
      [9] rpart_4.1.24           timechange_0.3.0       lifecycle_1.0.5        survival_3.8-3        
     [13] RSQLite_2.4.5          magrittr_2.0.4         compiler_4.5.2         rlang_1.1.7           
     [17] sass_0.4.10            tools_4.5.2            yaml_2.3.12            data.table_1.18.0     
-    [21] htmlwidgets_1.6.4      curl_7.0.0             bit_4.6.0              xml2_1.5.1            
+    [21] htmlwidgets_1.6.4      curl_7.0.0             bit_4.6.0              xml2_1.5.2            
     [25] RColorBrewer_1.1-3     desc_1.4.3             nnet_7.3-20            grid_4.5.2            
-    [29] xtable_1.8-4           e1071_1.7-17           future_1.68.0          ada_2.0-5             
[…]
