### 2026-01-27 c59fc6e
add "suwo_" preffix to all globally defined arguments
