+++ R/template_params.R
@@ -5,24 +5,24 @@
 #' some repositories might not support parallel queries from the same IP
 #' address as it might be identified as denial-of-service cyberattack.
 #' @param pb Logical argument to control if progress bar is shown. Default
-#' is \code{TRUE}. Can be set globally for the current R session via the "pb"
-#' option ( \code{options(pb = TRUE)}).
+#' is \code{TRUE}. Can be set globally for the current R session via the "suwo_pb"
+#' option ( \code{options(suwo_pb = TRUE)}).
 #' @param species Character string with the scientific name of a species in
 #' the format: "Genus epithet". Required. Can be set globally for the current
-#' R session via the "term" option
-#' (e.g. \code{options(term = "Hypsiboas rufitelus")}).
+#' R session via the "suwo_species" option
+#' (e.g. \code{options(suwo_species = "Hypsiboas rufitelus")}).
 #' @param verbose Logical argument that determines if text is shown in
 #' console. Default is \code{TRUE}. Can be set globally for the current R
-#' session via the "verbose" option
-#' ( \code{options(verbose = TRUE)}).
+#' session via the "suwo_verbose" option
+#' ( \code{options(suwo_verbose = TRUE)}).
 #' @param all_data Logical argument that determines if all data available
 #' from database is shown in the results of search. Default is \code{FALSE}.
-#' Can be set globally for the current R session via the "all_data"
-#' option ( \code{options(all_data = TRUE)}).
+#' Can be set globally for the current R session via the "suwo_all_data"
+#' option ( \code{options(suwo_all_data = TRUE)}).
 #' @param raw_data Logical argument that determines if the raw data from the
 #' repository is returned (e.g. without any manipulation).
 #' Default is \code{FALSE}. Can be set globally for the current R session
-#' via the "raw_data" option ( \code{options(raw_data = TRUE)}). If \code{TRUE}
+#' via the "suwo_raw_data" option ( \code{options(suwo_raw_data = TRUE)}). If \code{TRUE}
 #' \code{all_data} is set to \code{TRUE} internally. Useful for developers,
 #' or if users suspect that some data is mishandled during processing
 #' (i.e. date information is lost). Note that the metadata obtained
