### 2026-01-27 555ec9d
fix typos on table
