### 2026-01-27 f0abf12
remove importFrom when pkg::fun
