+++ R/merge_metadata.R
@@ -37,6 +37,7 @@
 #'
 #'  # combine metadata using a list of data frames
 #'  mt_list <- list(wikiaves = wa, gbif = gb)
+#'  merged_mt <- merge_metadata(mt_list)
 #' }
 #'
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
