+++ articles/suwo.md
@@ -91,13 +91,13 @@ metadata they retrieve:
 library(suwo)
 ```
 
-| Function                                                                             | Repository       | URL link                                                             | File types                                                              | Requires api key | Taxonomic level                           | Geographic coverage | Taxonomic coverage                                        | Other features                                        |
-|:-------------------------------------------------------------------------------------|:-----------------|:---------------------------------------------------------------------|:------------------------------------------------------------------------|:-----------------|:------------------------------------------|:--------------------|:----------------------------------------------------------|:------------------------------------------------------|
-| [query_gbif](https://marce10.github.io/suwo/reference/query_gbif.html)               | GBIF             | [https://www.gbif.org/](https://www.gbif.org/)                       | getOption, format, c(“image”, “sound”, “video”, “interactive resource”) | No               | Species                                   | Global              | All life                                                  | Specify query by data base                            |
-| [query_inaturalist](https://marce10.github.io/suwo/reference/query_inaturalist.html) | iNaturalist      | [https://www.inaturalist.org/](https://www.inaturalist.org/)         | getOption, format, c(“image”, “sound”)                                  | No               | Species                                   | Global              | All life                                                  |                                                       |
-| [query_macaulay](https://marce10.github.io/suwo/reference/query_macaulay.html)       | Macaulay Library | [https://www.macaulaylibrary.org/](https://www.macaulaylibrary.org/) | getOption, format, c(“image”, “sound”, “video”)                         | No               | Species                                   | Global              | Mostly birds but also other vertebrates and invertebrates | Interactive                                           |
-| [query_wikiaves](https://marce10.github.io/suwo/reference/query_wikiaves.html)       | WikiAves         | [https://www.wikiaves.com.br/](https://www.wikiaves.com.br/)         | getOption, format, c(“image”, “sound”)                                  | No               | Species                                   | Brazil              | Birds                                                     |                                                       |
-| [query_xenocanto](https://marce10.github.io/suwo/reference/query_xenocanto.html)     | Xeno-Canto       | [https://www.xeno-canto.org/](https://www.xeno-canto.org/)           | sound                                                                   | Yes              | Species, subspecies, genus, family, group | Global              | Birds, frogs, non-marine mammals and grasshoppers         | Specify query by taxonomy, geographic range and dates |
+| Function                                                                             | Repository       | URL link                                                             | File types                                | Requires api key | Taxonomic level                           | Geographic coverage | Taxonomic coverage                                        | Other features                                        |
[…]
