+++ R/query_inaturalist.R
@@ -7,12 +7,12 @@
 #' Currently 'image' and 'sound' are available. Can be set globally for
 #' the current R session via the "suwo_format" option
 #' (e.g. \code{options(suwo_format = "image")}). Required.
+#' @param all_data Logical argument that determines if all data available
+#' from database is shown in the results of search. Default is \code{TRUE}.
 #' @param identified Logical argument to define if search results are
 #' categorized as identified by inaturalist.
 #' @param verifiable Logical argument to define if search results are
 #' categorized as verifiable by inaturalist.
-#' @param all_data Logical argument that determines if all data available
-#' from database is shown in the results of search. Default is \code{TRUE}.
 #' @export
 #' @name query_inaturalist
 #' @return The function returns a data frame with the metadata of the media
@@ -37,14 +37,14 @@
 #'
 
 query_inaturalist <- function(species = getOption("suwo_species"),
-                            cores = getOption("mc.cores", 1),
-                            pb = getOption("suwo_pb", TRUE),
-                            verbose = getOption("suwo_verbose", TRUE),
-                             format = getOption("suwo_format", c("image", "sound")),
-                            identified = FALSE,
-                            verifiable = FALSE,
-                            all_data = getOption("suwo_all_data", FALSE),
-                            raw_data = getOption("suwo_raw_data", FALSE)) {
+                      format = getOption("suwo_format", c("image", "sound")),
+                      cores = getOption("mc.cores", 1),
+                      pb = getOption("suwo_pb", TRUE),
+                      verbose = getOption("suwo_verbose", TRUE),
+                      all_data = getOption("suwo_all_data", FALSE),
+                      raw_data = getOption("suwo_raw_data", FALSE),
+                      identified = FALSE,
+                      verifiable = FALSE) {
 
   arguments <- as.list(base::match.call())[-1]
   for (i in names(arguments)) arguments[[i]] <- get(i)
