### 2026-01-27 eec1c1c
replace .stop with cli::cli_abort
