### 2026-01-27 6d655b4
replace is with inherits
