### 2026-01-27 81fe5de
remove viridis
