+++ DESCRIPTION
@@ -14,7 +14,6 @@ Imports:
     methods,
     pbapply,
     utils,
-    viridis,
     rlang,
     httr2,
     jsonlite,
@@ -32,7 +31,7 @@ Suggests:
     rmarkdown,
     graphics
 Depends:
-    R (>= 3.5)
+    R (>= 4.0.0)
 Config/testthat/edition: 3
 RoxygenNote: 7.3.2
 Authors@R: 
