### 2026-01-27 3c3f29f
Update CITATION.cff
