+++ CITATION.cff
@@ -14,7 +14,7 @@ identifiers:
   value: https://github.com/maRce10/suwo
 abstract: Streamline searching/downloading of nature media files (e.g. audios, photos)
   from online repositories. The package offers functions for obtaining media metadata
-  from online repositories, downloading associated media files ad updating data sets
+  from online repositories, downloading associated media files and updating data sets
   with new records.
 authors:
 - family-names: Araya-Salas
@@ -62,7 +62,7 @@ references:
     given-names: Michel
     email: michellang@gmail.com
     orcid: https://orcid.org/0000-0001-9754-0393
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.checkmate
 - type: software
   title: cli
@@ -74,7 +74,7 @@ references:
   - family-names: Csárdi
     given-names: Gábor
     email: gabor@posit.co
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.cli
 - type: software
   title: methods
@@ -85,7 +85,7 @@ references:
   institution:
     name: R Foundation for Statistical Computing
     address: Vienna, Austria
-  year: '2025'
+  year: '2026'
 - type: software
   title: pbapply
   abstract: 'pbapply: Adding Progress Bar to ''*apply'' Functions'
@@ -100,7 +100,7 @@ references:
   - family-names: Zawadzki
     given-names: Zygmunt
     email: zygmunt@zstat.pl
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.pbapply
 - type: software
   title: utils
@@ -111,19 +111,7 @@ references:
   institution:
     name: R Foundation for Statistical Computing
     address: Vienna, Austria
-  year: '2025'
-- type: software
-  title: viridis
-  abstract: 'viridis: Colorblind-Friendly Color Maps for R'
-  notes: Imports
-  url: https://sjmgarnier.github.io/viridis/
-  repository: https://CRAN.R-project.org/package=viridis
-  authors:
-  - family-names: Garnier
-    given-names: Simon
-    email: garnier@njit.edu
-  year: '2025'
-  doi: 10.32614/CRAN.package.viridis
+  year: '2026'
 - type: software
   title: rlang
   abstract: 'rlang: Functions for Base Types and Core R and ''Tidyverse'' Features'
@@ -137,7 +125,7 @@ references:
   - family-names: Wickham
     given-names: Hadley
     email: hadley@posit.co
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.rlang
 - type: software
   title: httr2
@@ -149,7 +137,7 @@ references:
   - family-names: Wickham
     given-names: Hadley
     email: hadley@posit.co
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.httr2
 - type: software
   title: jsonlite
@@ -162,7 +150,7 @@ references:
     given-names: Jeroen
     email: jeroenooms@gmail.com
     orcid: https://orcid.org/0000-0002-4035-0289
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.jsonlite
 - type: software
   title: kableExtra
@@ -175,7 +163,7 @@ references:
     given-names: Hao
     email: haozhu233@gmail.com
     orcid: https://orcid.org/0000-0002-3386-6076
-  year: '2025'
+  year: '2026'
   doi: 10.32614/CRAN.package.kableExtra
 - type: software
   title: lubridate
@@ -191,7 +179,7 @@ references:
     given-names: Garrett
   - family-names: Wickham
[…]
