### 2026-01-28 0a2f2c4
fix api key and other small issues
