+++ reference/query_xenocanto.md
@@ -7,13 +7,13 @@
 
 ``` r
 query_xenocanto(
-  species = getOption("species"),
-  api_key = getOption("xc_api_key"),
+  species = getOption("suwo_species"),
   cores = getOption("mc.cores", 1),
-  pb = getOption("pb", TRUE),
-  verbose = getOption("verbose", TRUE),
-  all_data = getOption("all_data", FALSE),
-  raw_data = getOption("raw_data", FALSE)
+  pb = getOption("suwo_pb", TRUE),
+  verbose = getOption("suwo_verbose", TRUE),
+  all_data = getOption("suwo_all_data", FALSE),
+  raw_data = getOption("suwo_raw_data", FALSE),
+  api_key = Sys.getenv("xc_api_key")
 )
 ```
 
@@ -23,21 +23,15 @@ query_xenocanto(
 
   Character string with the scientific name of a species in the format:
   "Genus epithet". Required. Can be set globally for the current R
-  session via the "term" option (e.g.
-  `options(term = "Hypsiboas rufitelus")`). Alternatively, a character
-  string containing additional tags that follows the Xeno-Canto advanced
-  query syntax can be provided. Tags are of the form 'tag:searchterm'.
-  For instance, `'type:"song"'` will search for recordings where the
-  sound type contains 'song'. Multiple tags can be provided (e.g.,
-  `'"cnt:"belize" type:"song"'`). See examples down below and check
-  [Xeno-Canto's search help](https://www.xeno-canto.org/help/search) for
-  a full description.
-
-- api_key:
-
-  Character string refering to the key assigned by Xeno-Canto as
-  authorization for searches. Get yours at
-  <https://xeno-canto.org/account>. Required.
+  session via the "suwo_species" option (e.g.
+  `options(suwo_species = "Hypsiboas rufitelus")`). Alternatively, a
+  character string containing additional tags that follows the
+  Xeno-Canto advanced query syntax can be provided. Tags are of the form
+  'tag:searchterm'. For instance, `'type:"song"'` will search for
+  recordings where the sound type contains 'song'. Multiple tags can be
+  provided (e.g., `'"cnt:"belize" type:"song"'`). See examples down
+  below and check [Xeno-Canto's search
+  help](https://www.xeno-canto.org/help/search) for a full description.
 
 - cores:
 
@@ -52,32 +46,44 @@ query_xenocanto(
 - pb:
 
   Logical argument to control if progress bar is shown. Default is
-  `TRUE`. Can be set globally for the current R session via the "pb"
-  option ( `options(pb = TRUE)`).
+  `TRUE`. Can be set globally for the current R session via the
+  "suwo_pb" option ( `options(suwo_pb = TRUE)`).
 
 - verbose:
 
   Logical argument that determines if text is shown in console. Default
   is `TRUE`. Can be set globally for the current R session via the
-  "verbose" option ( `options(verbose = TRUE)`).
+  "suwo_verbose" option ( `options(suwo_verbose = TRUE)`).
 
 - all_data:
 
   Logical argument that determines if all data available from database
   is shown in the results of search. Default is `FALSE`. Can be set
-  globally for the current R session via the "all_data" option (
-  `options(all_data = TRUE)`).
+  globally for the current R session via the "suwo_all_data" option (
+  `options(suwo_all_data = TRUE)`).
 
 - raw_data:
 
   Logical argument that determines if the raw data from the repository
   is returned (e.g. without any manipulation). Default is `FALSE`. Can
-  be set globally for the current R session via the "raw_data" option (
-  `options(raw_data = TRUE)`). If `TRUE` `all_data` is set to `TRUE`
[…]
