+++ reference/query_gbif.md
@@ -6,14 +6,14 @@
 
 ``` r
 query_gbif(
-  species = getOption("species"),
-  format = getOption("format", c("sound", "image", "video", "interactive resource")),
+  species = getOption("suwo_species"),
+  format = getOption("suwo_format", c("image", "sound", "video", "interactive resource")),
   cores = getOption("mc.cores", 1),
-  pb = getOption("pb", TRUE),
-  verbose = getOption("verbose", TRUE),
+  pb = getOption("suwo_pb", TRUE),
+  verbose = getOption("suwo_verbose", TRUE),
   dataset = NULL,
-  all_data = getOption("all_data", FALSE),
-  raw_data = getOption("raw_data", FALSE)
+  all_data = getOption("suwo_all_data", FALSE),
+  raw_data = getOption("suwo_raw_data", FALSE)
 )
 ```
 
@@ -23,15 +23,15 @@ query_gbif(
 
   Character string with the scientific name of a species in the format:
   "Genus epithet". Required. Can be set globally for the current R
-  session via the "term" option (e.g.
-  `options(term = "Hypsiboas rufitelus")`).
+  session via the "suwo_species" option (e.g.
+  `options(suwo_species = "Hypsiboas rufitelus")`).
 
 - format:
 
   Character vector with the media format to query for. Options are
   'sound', 'image', 'video' and 'interactive resource'. Can be set
-  globally for the current R session via the "format" option (e.g.
-  `options(format = "sound")`). Required.
+  globally for the current R session via the "suwo_format" option (e.g.
+  `options(suwo_format = "image")`). Required.
 
 - cores:
 
@@ -46,14 +46,14 @@ query_gbif(
 - pb:
 
   Logical argument to control if progress bar is shown. Default is
-  `TRUE`. Can be set globally for the current R session via the "pb"
-  option ( `options(pb = TRUE)`).
+  `TRUE`. Can be set globally for the current R session via the
+  "suwo_pb" option ( `options(suwo_pb = TRUE)`).
 
 - verbose:
 
   Logical argument that determines if text is shown in console. Default
   is `TRUE`. Can be set globally for the current R session via the
-  "verbose" option ( `options(verbose = TRUE)`).
+  "suwo_verbose" option ( `options(suwo_verbose = TRUE)`).
 
 - dataset:
 
@@ -67,19 +67,20 @@ query_gbif(
 
   Logical argument that determines if all data available from database
   is shown in the results of search. Default is `FALSE`. Can be set
-  globally for the current R session via the "all_data" option (
-  `options(all_data = TRUE)`).
+  globally for the current R session via the "suwo_all_data" option (
+  `options(suwo_all_data = TRUE)`).
 
 - raw_data:
 
   Logical argument that determines if the raw data from the repository
   is returned (e.g. without any manipulation). Default is `FALSE`. Can
-  be set globally for the current R session via the "raw_data" option (
-  `options(raw_data = TRUE)`). If `TRUE` `all_data` is set to `TRUE`
-  internally. Useful for developers, or if users suspect that some data
-  is mishandled during processing (i.e. date information is lost). Note
-  that the metadata obtained when `raw_data = TRUE` is not standardized,
-  so most suwo functions for downstream steps will not work on them.
+  be set globally for the current R session via the "suwo_raw_data"
+  option ( `options(suwo_raw_data = TRUE)`). If `TRUE` `all_data` is set
+  to `TRUE` internally. Useful for developers, or if users suspect that
+  some data is mishandled during processing (i.e. date information is
+  lost). Note that the metadata obtained when `raw_data = TRUE` is not
+  standardized, so most suwo functions for downstream steps will not
+  work on them.
 
 ## Value
 
