### 2026-01-31 acc3d03
Updated readme for commercial use
