### 2026-02-04 fcedb4c
error message when missing "duplicated_group" column in remove_duplicates()
