+++ R/query_gbif.R
@@ -47,15 +47,15 @@ query_gbif <-
            all_data = getOption("suwo_all_data", FALSE),
            raw_data = getOption("suwo_raw_data", FALSE)) {
     # check arguments
-    arguments <- as.list(base::match.call())[-1]
+    arguments <- as.list(base::match.call())
 
     # add objects to argument names
-    for (i in names(arguments)) {
+   for (i in names(arguments)[-1]) {
       arguments[[i]] <- get(i)
     }
 
     # check each arguments
-    check_results <- .check_arguments(args = arguments)
+    check_results <- .check_arguments(fun = arguments[[1]], args = arguments)
 
     # report errors
     checkmate::reportAssertions(check_results)
