+++ R/query_wikiaves.R
@@ -42,15 +42,15 @@ query_wikiaves <-
            all_data = getOption("suwo_all_data", FALSE),
            raw_data = getOption("suwo_raw_data", FALSE)) {
     # check arguments
-    arguments <- as.list(base::match.call())[-1]
+    arguments <- as.list(base::match.call())
 
     # add objects to argument names
-    for (i in names(arguments)) {
+   for (i in names(arguments)[-1]) {
       arguments[[i]] <- get(i)
     }
 
     # check each arguments
-    check_results <- .check_arguments(args = arguments)
+    check_results <- .check_arguments(fun = arguments[[1]], args = arguments)
 
     # report errors
     checkmate::reportAssertions(check_results)
