+++ tests/testthat/test_remove_duplicates.R
@@ -3,8 +3,6 @@ options(verbose = TRUE)
 
 test_that("find from gbif ml and xc", {
 
-  skip_on_cran()
-  skip_if_offline()
   # find duplicates
   label_dup_metadata <- find_duplicates(metadata = suwo:::merged_metadata)
 
@@ -20,3 +18,20 @@ test_that("find from gbif ml and xc", {
 
   })
 
+
+test_that("error if missing colum", {
+
+  # find duplicates
+  label_dup_metadata <- find_duplicates(metadata = suwo:::merged_metadata)
+
+  # remove duplicates
+  dedup_metadata <- remove_duplicates(suwo:::merged_metadata)
+
+  expect_true(nrow(dedup_metadata) > 2026)
+
+  # remove duplicates
+  dedup_metadata2 <- remove_duplicates(label_dup_metadata, same_repo = TRUE)
+
+  expect_true(nrow(dedup_metadata2) > 1891)
+
+})
