### 2026-02-04 e7a966f
further explain duplicate group
