+++ R/remove_duplicates.R
@@ -1,13 +1,13 @@
 #' Remove duplicated media records
 #'
-#' \code{remove_duplicates} removes duplicated media records.
+#' `remove_duplicates` removes duplicated media records.
 #' @inheritParams template_params
 #' @param metadata data frame obtained from possible duplicates  with the
-#' function \code{\link{find_duplicates}}. The data frame must have the column
-#' 'duplicate_group' returned by \code{\link{find_duplicates}}.
+#' function [find_duplicates()]. The data frame must have the column
+#' 'duplicate_group' returned by [find_duplicates()].
 #' @param same_repo Logical argument indicating if observations labeled
 #' as duplicates that belong to the same repository should be removed. Default
-#' is \code{FALSE}. If \code{TRUE}, only one of the duplicated observations
+#' is `FALSE`. If `TRUE`, only one of the duplicated observations
 #' from the same repository will be retained in the output data frame. This is
 #' useful as it can be expected that observations from the same repository are
 #' not true duplicates (e.g. different recordings uploaded to Xeno-Canto with
@@ -23,17 +23,23 @@
 #' observations that were determined not to be duplicates.
 #' @export
 #' @name remove_duplicates
-#' @details This function removes duplicate observations identified with the
-#' function \code{\link{find_duplicates}}. When duplicates are found, one
-#' observation from each group of duplicates is retained in the output data
-#' frame. However, if multiple observations from the same repository are
-#' labeled as duplicates, by default (\code{same_repo = FALSE}) all of them
+#' @details When compiling data from multiple repositories, duplicated media
+#' records are a common issue, particularly for sound recordings. These
+#' duplicates occur both through data sharing between repositories like
+#' Xeno-Canto and GBIF, and when users upload the same file to multiple
+#' platforms. In such cases those multiple observations seem to refer to the same
+#' media file and therefore, only one copy is needed. This function removes
+#' duplicate observations identified with the function
+#' [find_duplicates()]. When duplicates are found, one observation
+#' from each group of duplicates is retained in the output data frame.
+#' However, if multiple observations from the same repository are
+#' labeled as duplicates, by default (`same_repo = FALSE`) all of them
 #' are retained in the output data frame. This is useful as it can be
 #' expected that observations from the same repository are not true
 #' duplicates (e.g. different recordings uploaded to Xeno-Canto with
 #' the same date, time and location by the same user), but rather have not
 #' been documented with enough precision to be told apart. This behavior can
-#' be modified. If \code{same_repo = TRUE}, only one of the duplicated
+#' be modified. If `same_repo = TRUE`, only one of the duplicated
 #' observations from the same repository will be retained in the output data
 #' frame. The function will give priority to repositories in which media
 #' downloading is more straightforward (Xeno-Canto and GBIF), but this can be
@@ -56,7 +62,7 @@
 #' dedup_metadata <- remove_duplicates(label_dup_metadata)
 #' }
 #'
-#' @seealso \code{\link{find_duplicates}}, \code{\link{merge_metadata}}
+#' @seealso [find_duplicates()], [merge_metadata()]
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
 #'
 remove_duplicates <-
