### 2026-02-05 cd2aaa9
Revert "Updated alt text for flow diagram"
This reverts commit 9908b1579220c46ed1112a78306023cca6d5fc8e.
