### 2026-02-05 a71a1da
Revert "Merge branch 'main' of https://github.com/maRce10/suwo"
This reverts commit a1a5154cc76bada88ce656880e4e63e3cb2f360a, reversing
changes made to 9908b1579220c46ed1112a78306023cca6d5fc8e.
