+++ articles/suwo.md
@@ -665,7 +665,7 @@ head(azam_files, 4)
 fs::dir_tree(path = out_folder)
 ```
 
-     /tmp/RtmpbrD1tk/amanita_zambiana 
+     /tmp/Rtmpw4Mwx9/amanita_zambiana 
     ├──  Amanita_zambiana-GBIF3759537817-1.jpeg 
     ├──  Amanita_zambiana-GBIF3759537817-2.jpeg 
     ├──  Amanita_zambiana-GBIF4430877067-1.jpeg 
@@ -752,7 +752,7 @@ dhol_files <- download_media(metadata = d_holocanthus,
 fs::dir_tree(path = out_folder)
 ```
 
-     /tmp/RtmpbrD1tk/diodon_holocanthus 
+     /tmp/Rtmpw4Mwx9/diodon_holocanthus 
     ├──  Cabo Verde 
     │   └──  Diodon_holocanthus-GBIF3985886532.jpeg 
     ├──  Cayman Islands 
@@ -835,24 +835,24 @@ Click to see
     [1] kableExtra_1.4.0 suwo_0.1.0       knitr_1.51      
 
     loaded via a namespace (and not attached):
-     [1] viridisLite_0.4.2      farver_2.1.2           blob_1.3.0             fastmap_1.2.0         
+     [1] viridisLite_0.4.3      farver_2.1.2           blob_1.3.0             fastmap_1.2.0         
      [5] digest_0.6.39          rpart_4.1.24           timechange_0.4.0       lifecycle_1.0.5       
      [9] survival_3.8-3         RSQLite_2.4.5          magrittr_2.0.4         compiler_4.5.2        
     [13] rlang_1.1.7            sass_0.4.10            tools_4.5.2            yaml_2.3.12           
     [17] data.table_1.18.2.1    htmlwidgets_1.6.4      bit_4.6.0              curl_7.0.0            
     [21] xml2_1.5.2             RColorBrewer_1.1-3     desc_1.4.3             nnet_7.3-20           
     [25] grid_4.5.2             xtable_1.8-4           e1071_1.7-17           future_1.69.0         
-    [29] globals_0.18.0         scales_1.4.0           MASS_7.3-65            cli_3.6.5             
+    [29] globals_0.19.0         scales_1.4.0           MASS_7.3-65            cli_3.6.5             
     [33] rmarkdown_2.30         crayon_1.5.3           ragg_1.5.0             generics_0.1.4        
     [37] rstudioapi_0.18.0      RecordLinkage_0.4-12.6 future.apply_1.20.1    DBI_1.2.3             
     [41] pbapply_1.7-4          cachem_1.1.0           proxy_0.4-29           stringr_1.6.0         
     [45] splines_4.5.2          parallel_4.5.2         vctrs_0.7.1            Matrix_1.7-4          
     [49] jsonlite_2.0.0         bit64_4.6.0-1          listenv_0.10.0         jpeg_0.1-11           
     [53] systemfonts_1.3.1      evd_2.3-7.1            jquerylib_0.1.4        glue_1.8.0            
-    [57] parallelly_1.46.1      pkgdown_2.2.0          codetools_0.2-20       lubridate_1.9.4       
+    [57] parallelly_1.46.1      pkgdown_2.2.0          codetools_0.2-20       lubridate_1.9.5       
     [61] stringi_1.8.7          tibble_3.3.1           pillar_1.11.1          rappdirs_0.3.4        
     [65] htmltools_0.5.9        ipred_0.9-15           lava_1.8.2             R6_2.6.1              
     [69] ff_4.5.2               httr2_1.2.2            textshaping_1.0.4      evaluate_1.0.5        
[…]
