### 2026-02-07 1d5ce6e
added cli progress bar with parallel
