+++ R/template_params.R
@@ -6,7 +6,8 @@
 #' address as it might be identified as denial-of-service cyberattack.
 #' @param pb Logical argument to control if progress bar is shown. Default
 #' is `TRUE`. Can be set globally for the current R session via the "suwo_pb"
-#' option ( `options(suwo_pb = TRUE)`).
+#' option ( `options(suwo_pb = TRUE)`). Not shown if only a few observations
+#' are found.
 #' @param species Character string with the scientific name of a species in
 #' the format: "Genus epithet". Required. Can be set globally for the current
 #' R session via the "suwo_species" option
