+++ R/query_wikiaves.R
@@ -34,18 +34,20 @@
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
 
 query_wikiaves <-
-  function(species = getOption("suwo_species"),
-           format = getOption("suwo_format", c("image", "sound")),
-           cores = getOption("mc.cores", 1),
-           pb = getOption("suwo_pb", TRUE),
-           verbose = getOption("suwo_verbose", TRUE),
-           all_data = getOption("suwo_all_data", FALSE),
-           raw_data = getOption("suwo_raw_data", FALSE)) {
+  function(
+    species = getOption("suwo_species"),
+    format = getOption("suwo_format", c("image", "sound")),
+    cores = getOption("mc.cores", 1),
+    pb = getOption("suwo_pb", TRUE),
+    verbose = getOption("suwo_verbose", TRUE),
+    all_data = getOption("suwo_all_data", FALSE),
+    raw_data = getOption("suwo_raw_data", FALSE)
+  ) {
     # check arguments
     arguments <- as.list(base::match.call())
 
     # add objects to argument names
-   for (i in names(arguments)[-1]) {
+    for (i in names(arguments)[-1]) {
       arguments[[i]] <- get(i)
     }
 
@@ -72,8 +74,10 @@ query_wikiaves <-
         gsub(" ", "%20", species)
       )
     )
-    request_obj <- httr2::req_user_agent(request_obj,
-                                      "suwo (https://github.com/maRce10/suwo)")
+    request_obj <- httr2::req_user_agent(
+      request_obj,
+      "suwo (https://github.com/maRce10/suwo)"
+    )
     response <- httr2::req_perform(request_obj)
 
     # check if request succeeded
@@ -99,42 +103,53 @@ query_wikiaves <-
       return(invisible(NULL))
     }
 
-    get_ids$total_registers <- vapply(seq_len(nrow(get_ids)), function(u) {
-      request_obj <- httr2::request(
-      paste0(
-          "https://www.wikiaves.com.br/getRegistrosJSON.php?tm=",
-          wiki_format,
-          "&t=s&s=",
-          get_ids$id[u],
-          "&o=mp&p=1"
+    get_ids$total_registers <- vapply(
+      seq_len(nrow(get_ids)),
+      function(u) {
+        request_obj <- httr2::request(
+          paste0(
+            "https://www.wikiaves.com.br/getRegistrosJSON.php?tm=",
+            wiki_format,
+            "&t=s&s=",
+            get_ids$id[u],
+            "&o=mp&p=1"
+          )
         )
-      )
-      request_obj <-
-        httr2::req_user_agent(request_obj,
-              "suwo (https://github.com/maRce10/suwo)")
-      request_obj <- httr2::req_error(request_obj,
-                                      is_error = function(resp) FALSE)
+        request_obj <-
+          httr2::req_user_agent(
+            request_obj,
+            "suwo (https://github.com/maRce10/suwo)"
+          )
+        request_obj <- httr2::req_error(request_obj, is_error = function(resp) {
+          FALSE
+        })
 
-      response <- try(httr2::req_perform(request_obj), silent = TRUE)
+        response <- try(httr2::req_perform(request_obj), silent = TRUE)
 
-      # if fail request return -9999
-      if (.is_error(response)) {
-        return(-999)
-      }
+        # if fail request return -9999
+        if (.is_error(response)) {
+          return(-999)
+        }
 
-      # check if request succeeded
-      if (httr2::resp_is_error(response)) {
-        return(-999)
-      }
+        # check if request succeeded
+        if (httr2::resp_is_error(response)) {
+          return(-999)
+        }
 
-      content <- httr2::resp_body_json(response)
-      as.numeric(content$registros$total)
-
-    }, numeric(1))
[…]
