+++ data-raw/DATASET.R
@@ -1,19 +1,17 @@
 ## code to prepare `DATASET` dataset goes here
 # see https://r-pkgs.org/data.html#sec-data-sysdata
 
-
 # Taxon code for macaulay
 # downloaded from https://www.birds.cornell.edu/clementschecklist/introduction
 #/updateindex/october-2025/2025-citation-checklist-downloads/
-ml_taxon_code <- utils::read.csv("./examples/eBird_taxonomy_v2025.csv",
-                          na.strings = "")
+ml_taxon_code <- utils::read.csv(
+  "./examples/eBird_taxonomy_v2025.csv",
+  na.strings = ""
+)
 
 
-xc_adf <-  query_xenocanto(all_data = FALSE,
-                           api_key = Sys.getenv("xc_api_key"))
-gb_adf_s <- query_gbif(species = term,
-                       format = "sound",
-                       all_data = FALSE)
+xc_adf <- query_xenocanto(all_data = FALSE, api_key = Sys.getenv("xc_api_key"))
+gb_adf_s <- query_gbif(species = term, format = "sound", all_data = FALSE)
 ml_adf_s <- query_macaulay(
   species = term,
   format = "sound",
@@ -25,25 +23,31 @@ merged_metadata <- merge_metadata(xc_adf, gb_adf_s, ml_adf_s)
 
 options(all_data = TRUE)
 
-h_sarapiquensis <- head(query_inaturalist(species = "Heliconia sarapiquensis",
-                                          format = "image"),
-                        4)
+h_sarapiquensis <- head(
+  query_inaturalist(species = "Heliconia sarapiquensis", format = "image"),
+  4
+)
 h_harpyja <- head(query_wikiaves(species = "Harpia harpyja"), 4)
 a_hahneli <- query_xenocanto(
   species = 'sp:"Ameerega hahneli" cnt:"French Guiana" q:"A"',
-                             api_key = Sys.getenv("xc_api_key"))
-p_lotor <- head(query_gbif(species = "Procyon lotor", format = "video"), 4)
-t_rufiventris <- head(query_macaulay(
-  species = "Turdus rufiventris",
-  format = "sound",
-  path = tempdir()
-),
-4)
-t_tricolor <- head(query_xenocanto(
-  species = "Thyroptera tricolor",
   api_key = Sys.getenv("xc_api_key")
-),
-4)
+)
+p_lotor <- head(query_gbif(species = "Procyon lotor", format = "video"), 4)
+t_rufiventris <- head(
+  query_macaulay(
+    species = "Turdus rufiventris",
+    format = "sound",
+    path = tempdir()
+  ),
+  4
+)
+t_tricolor <- head(
+  query_xenocanto(
+    species = "Thyroptera tricolor",
+    api_key = Sys.getenv("xc_api_key")
+  ),
+  4
+)
 
 p_striigularis <- head(
   query_macaulay(
