### 2026-02-09 6e93735
seq_len(length()) replace with seq_along()
