+++ R/download_media.R
@@ -99,7 +99,7 @@ download_media <-
     metadata$non_dup_key <- unlist(lapply(unique(metadata$key), function(x) {
       on <- metadata$key[metadata$key == x]
       if (length(on) > 1) {
-        return(paste0(on, "-", seq_len(length(on))))
+        return(paste0(on, "-", seq_along(on)))
       } else {
         return(x)
       }
