### 2026-02-09 3d0d646
use httr2::resp_body_json().
