+++ R/query_gbif.R
@@ -37,6 +37,7 @@
 #'
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
 #'
+
 query_gbif <-
   function(
     species = getOption("suwo_species"),
@@ -51,163 +52,152 @@ query_gbif <-
     all_data = getOption("suwo_all_data", FALSE),
     raw_data = getOption("suwo_raw_data", FALSE)
   ) {
-    # check arguments
+    ##  argument checking
     arguments <- as.list(base::match.call())
-
-    # add objects to argument names
     for (i in names(arguments)[-1]) {
       arguments[[i]] <- get(i)
     }
-
-    # check each arguments
     check_results <- .check_arguments(fun = arguments[[1]], args = arguments)
-
-    # report errors
     checkmate::reportAssertions(check_results)
 
-    # Use the unified connection checker
+    ##  connection
     if (!.checkconnection(verb = verbose, service = "gbif")) {
       return(invisible(NULL))
     }
 
-    # assign a value to format
+    ##  format
     format <- rlang::arg_match(format)
 
     gbif_format <- switch(
       format,
       sound = "Sound",
-      `image` = "StillImage",
-      `video` = "MovingImage",
+      image = "StillImage",
+      video = "MovingImage",
       `interactive resource` = "InteractiveResource"
     )
 
-    srch_trm <- paste0(
-      "https://api.gbif.org/v1/occurrence/search?limit=300&",
-      if (is.null(dataset)) {
-        ""
-      } else {
-        "datasetKey="
-      },
-      dataset,
-      "&scientificName=",
-      gsub(" ", "%20", species),
-      "&media_type=",
-      gbif_format
+    ##  base request
+    base_request <-
+      httr2::request("https://api.gbif.org/v1/occurrence/search") |>
+      httr2::req_user_agent("suwo (https://github.com/maRce10/suwo)") |>
+      httr2::req_url_query(
+        limit = 300,
+        scientificName = species,
+        media_type = gbif_format,
+        datasetKey = dataset
+      ) |>
+      httr2::req_error(is_error = function(resp) FALSE)
+
+    ##  first request (metadata)
+    response <- try(httr2::req_perform(base_request), silent = TRUE)
+
+    if (.is_error(response) || httr2::resp_is_error(response)) {
+      if (verbose) {
+        .message("Metadata could not be downloaded", as = "failure")
+      }
+      return(invisible(NULL))
+    }
+
+    base.srch.pth <- httr2::resp_body_json(
+      response,
+      simplifyVector = TRUE
     )
 
-    base.srch.pth <- try(jsonlite::fromJSON(srch_trm), silent = TRUE)
-
-    # let user know error when downloading metadata
-    if (.is_error(base.srch.pth)) {
-      if (verbose) {
-        .message(text = "Metadata could not be downloaded", as = "failure")
-      }
-      return(invisible(NULL))
-    }
-
-    # message if nothing found
     if (base.srch.pth$count == 0) {
       if (verbose) {
-        .message(text = "No matching records found", as = "failure")
+        .message("No matching records found", as = "failure")
       }
       return(invisible(NULL))
     }
 
-    # message number of results
     if (verbose) {
       .message(n = base.srch.pth$count, as = "success")
     }
 
-    # get total number of pages
-    offsets <- (seq_len(ceiling(
-      base.srch.pth$count / base.srch.pth$limit
-    )) -
+    ##  pagination
+    offsets <- (seq_len(
+      ceiling(base.srch.pth$count / base.srch.pth$limit)
+    ) -
       1) *
-      300
+      base.srch.pth$limit
 
+    ##  paged download
[…]
