### 2026-02-09 4385598
remove get from argument checking
