+++ R/update_metadata.R
@@ -60,16 +60,19 @@ update_metadata <-
     api_key = NULL,
     dates = NULL
   ) {
-    # check arguments
-    arguments <- as.list(base::match.call())
-
-    # add objects to argument names
-    for (i in names(arguments)[-1]) {
-      arguments[[i]] <- get(i)
-    }
-
-    # check each arguments
-    check_results <- .check_arguments(fun = arguments[[1]], args = arguments)
+    ##  argument checking
+    check_results <- .check_arguments(
+      fun = "query_gbif",
+      args = list(
+        metadata = metadata,
+        path = path,
+        cores = cores,
+        pb = pb,
+        verbose = verbose,
+        api_key = api_key,
+        dates = dates
+      )
+    )
 
     # report errors
     checkmate::reportAssertions(check_results)
