+++ R/query_inaturalist.R
@@ -47,213 +47,22 @@ query_inaturalist <- function(
   identified = FALSE,
   verifiable = FALSE
 ) {
-  arguments <- as.list(base::match.call())
-  for (i in names(arguments)[-1]) {
-    arguments[[i]] <- get(i)
-  }
-
-  check_results <- .check_arguments(fun = arguments[[1]], args = arguments)
-  checkmate::reportAssertions(check_results)
-
-  format <- rlang::arg_match(format)
-  inat_format <- switch(format, sound = "sounds", image = "photos")
-
-  if (!.checkconnection(verb = verbose, service = "inat")) {
-    return(invisible(NULL))
-  }
-
-  base_url <- paste0(
-    "https://api.inaturalist.org/v1/observations?",
-    "per_page=200&",
-    "taxon_name=",
-    gsub(" ", "%20", species),
-    "&",
-    inat_format,
-    "=true&",
-    "identified=",
-    identified,
-    "&",
-    "verifiable=",
-    verifiable
-  )
-
-  first_query <- try(jsonlite::fromJSON(base_url), silent = TRUE)
-
-  if (.is_error(first_query) || is.null(first_query$total_results)) {
-    if (verbose) {
-      .message(text = "Metadata could not be downloaded", as = "failure")
-    }
-    return(invisible(NULL))
-  }
-
-  total_results <- first_query$total_results
-
-  if (total_results == 0) {
-    if (verbose) {
-      .message(text = "No matching records found", as = "failure")
-    }
-    return(invisible(NULL))
-  }
-
-  if (verbose) {
-    .message(n = total_results, as = "success")
-  }
-
-  pages <- seq_len(ceiling(total_results / 200))
-
-  if (Sys.info()[1] == "Windows" && cores > 1) {
-    cl <- parallel::makePSOCKcluster(cores)
-  } else {
-    cl <- cores
-  }
-
-  query_output_list <- .pbapply_sw(
-    X = pages,
-    cl = cl,
-    pbar = pb,
-    FUN = function(x, Y = pages) {
-      # set index to get the right page
-      page <- Y[x]
-
-      query_output <- try(
-        jsonlite::fromJSON(paste0(base_url, "&page=", page)),
-        silent = TRUE
-      )
-
-      if (.is_error(query_output) || is.null(query_output$results)) {
-        return(NULL)
-      }
-
-      query_output$results <- lapply(
-        seq_len(nrow(query_output$results)),
-        function(u) {
-          x <- as.data.frame(query_output$results[u, ])
-
-          media_df <- if (format == "sound") {
-            do.call(rbind, x$sounds)
-          } else {
-            do.call(rbind, x$photos)
-          }
-
-          media_df <- media_df[!vapply(media_df, is.list, logical(1))]
-          media_df <- data.frame(media_df)
-
-          x <- x[!vapply(x, is.list, logical(1))]
-          X_df <- data.frame(t(unlist(x)))
-          X_df <- cbind(X_df, media_df)
-
-          # modify URL so it refers to original file
-          if (format == "image") {
-            X_df$url <- gsub("square", "original", X_df$url)
-          }
-
-          return(X_df)
-        }
-      )
-
-      .merge_data_frames(query_output$results)
-    }
-  )
-
-  if (any(vapply(query_output_list, .is_error, logical(1)))) {
-    if (verbose) {
-      .message(text = "Metadata could not be downloaded", as = "failure")
-    }
-    return(invisible(NULL))
-  }
-
-  query_output_df <- .merge_data_frames(query_output_list)
-
-  split_location <- do.call(
-    rbind,
-    strsplit(as.character(query_output_df$location), ",")
-  )
-
[…]
