+++ R/query_xenocanto.R
@@ -74,11 +74,18 @@ query_xenocanto <-
     raw_data = getOption("suwo_raw_data", FALSE),
     api_key = Sys.getenv("xc_api_key")
   ) {
-    # check arguments
-    arguments <- as.list(base::match.call())
-    for (i in names(arguments)[-1]) {
-      arguments[[i]] <- get(i)
-    }
+    ##  argument checking
+    check_results <- .check_arguments(
+      fun = "query_xenocanto",
+      args = list(
+        species = species,
+        cores = cores,
+        pb = pb,
+        verbose = verbose,
+        all_data = all_data,
+        raw_data = raw_data
+      )
+    )
 
     # Check for API key
     if (is.null(api_key) || !nzchar(api_key)) {
