### 2026-02-10 8d49a77
precompile vignette
