+++ tests/testthat/test_query_wikiaves.R
@@ -4,9 +4,12 @@ test_that("search Glaucis dohrnii sound", {
   skip_on_cran()
   skip_if_offline()
 
-  df1 <- query_wikiaves(species = 'Glaucis dohrnii', format = "sound")
+  df1 <- try(
+    query_wikiaves(species = 'Glaucis dohrnii', format = "sound"),
+    silent = TRUE
+  )
 
-  skip_if(is.null(df1))
+  skip_if(suwo:::.is_error(df1))
 
   expect_true(nrow(df1) >= 30)
 })
@@ -16,9 +19,12 @@ test_that("search Glaucis dohrnii photos", {
   skip_on_cran()
   skip_if_offline()
 
-  df1 <- query_wikiaves(species = 'Glaucis dohrnii', format = "image")
+  df1 <- try(
+    query_wikiaves(species = 'Glaucis dohrnii', format = "image"),
+    silent = TRUE
+  )
 
-  skip_if(is.null(df1))
+  skip_if(suwo:::.is_error(df1))
 
   expect_true(nrow(df1) >= 420)
 })
@@ -28,9 +34,7 @@ test_that("no result", {
   skip_on_cran()
   skip_if_offline()
 
-  df1 <- query_wikiaves(species = 'asdasdasd', format = "image")
-
-  expect_true(is.null(df1))
+  expect_null(query_wikiaves(species = 'asdasdasd', format = "image"))
 })
 
 
@@ -38,14 +42,17 @@ test_that("test verbose FALSE", {
   skip_on_cran()
   skip_if_offline()
 
-  df1 <- testthat::capture_output(query_wikiaves(
-    species = 'a3',
-    format = "sound",
-    verbose = FALSE,
-    pb = FALSE
-  ))
+  df1 <- try(
+    testthat::capture_output(query_wikiaves(
+      species = 'a3',
+      format = "sound",
+      verbose = FALSE,
+      pb = FALSE
+    )),
+    silent = TRUE
+  )
 
-  skip_if(is.null(df1))
+  skip_if(suwo:::.is_error(df1))
 
   expect_true(df1 == "")
 })
@@ -54,15 +61,18 @@ test_that("test all_data FALSE", {
   skip_on_cran()
   skip_if_offline()
 
-  df1 <- query_wikiaves(
-    species = 'Glaucis dohrnii',
-    format = "image",
-    all_data = FALSE
+  df1 <- try(
+    query_wikiaves(
+      species = 'Glaucis dohrnii',
+      format = "sound",
+      all_data = FALSE
+    ),
+    silent = TRUE
   )
 
-  skip_if(is.null(df1))
+  skip_if(suwo:::.is_error(df1))
 
-  expected_col_names <- .format_query_output(only_basic_columns = T)
+  expected_col_names <- suwo:::.format_query_output(only_basic_columns = T)
 
   query_col_names <- colnames(df1)
 
