+++ R/query_wikiaves.R
@@ -73,9 +73,17 @@ query_wikiaves <-
     # initialize search with user agent
     request_obj <- httr2::request(
       "https://www.wikiaves.com.br/getTaxonsJSON.php"
-    ) |>
-      httr2::req_url_query(term = species) |>
-      httr2::req_user_agent("suwo (https://github.com/maRce10/suwo)")
+    )
+
+    request_obj <- httr2::req_url_query(
+      request_obj,
+      term = species
+    )
+
+    request_obj <- httr2::req_user_agent(
+      request_obj,
+      "suwo (https://github.com/maRce10/suwo)"
+    )
 
     response <- httr2::req_perform(request_obj)
 
