+++ tests/testthat/test_query_wikiaves.R
@@ -71,6 +71,7 @@ test_that("test all_data FALSE", {
   )
 
   skip_if(suwo:::.is_error(df1))
+  skip_if(is.null(df1))
 
   expected_col_names <- suwo:::.format_query_output(only_basic_columns = T)
 
