+++ data-raw/DATASET.R
@@ -1,7 +1,7 @@
 ## code to prepare `DATASET` dataset goes here
 # see https://r-pkgs.org/data.html#sec-data-sysdata
 
-# Taxon code for macaulay
+### Taxon code for macaulay ####
 # downloaded from https://www.birds.cornell.edu/clementschecklist/introduction
 #/updateindex/october-2025/2025-citation-checklist-downloads/
 ml_taxon_code <- utils::read.csv(
@@ -9,19 +9,24 @@ ml_taxon_code <- utils::read.csv(
   na.strings = ""
 )
 
+### CREATING METADATA TESTING ####
+### example merged metadata ####
 
-xc_adf <- query_xenocanto(all_data = FALSE, api_key = Sys.getenv("xc_api_key"))
-gb_adf_s <- query_gbif(species = term, format = "sound", all_data = FALSE)
-ml_adf_s <- query_macaulay(
-  species = term,
-  format = "sound",
-  path = tempdir(),
-  all_data = FALSE
+options(
+  suwo_species = "Turdus rufiventris",
+  suwo_format = "sound",
+  suwo_all_data = FALSE
 )
+
+xc_adf <- query_xenocanto()
+gb_adf_s <- query_gbif()
+ml_adf_s <- query_macaulay(path = tempdir())
 
 merged_metadata <- merge_metadata(xc_adf, gb_adf_s, ml_adf_s)
 
-options(all_data = TRUE)
+
+### CREATE ADDITIONAL METADATA FOR TESTING ####
+options(suwo_all_data = TRUE)
 
 h_sarapiquensis <- head(
   query_inaturalist(species = "Heliconia sarapiquensis", format = "image"),
@@ -67,8 +72,7 @@ d_holocanthus <- d_holocanthus[d_holocanthus$file_extension == "jpeg", ]
 set.seed(666)
 d_holocanthus <- d_holocanthus[sample(seq_len(nrow(d_holocanthus)), 6), ]
 
-vignette_metadata <- list(
-  h_wagneriana = h_wagneriana,
+testing_metadata <- list(
   h_harpyja = h_harpyja,
   a_hahneli = a_hahneli,
   p_lotor = p_lotor,
@@ -82,7 +86,166 @@ vignette_metadata <- list(
 usethis::use_data(
   ml_taxon_code,
   merged_metadata,
-  vignette_metadata,
+  testing_metadata,
   internal = TRUE,
   overwrite = TRUE
+)
+
+
+### create function description table for vignette
+metadata_list <- testing_metadata
+
+Repository <- vapply(
+  metadata_list,
+  function(x) {
+    x[1, "repository"]
+  },
+  character(1)
+)
+
+Repository[Repository == "Wikiaves"] <- "WikiAves"
+
+names(metadata_list) <- tolower(Repository)
+
+Function <- c(
+  gbif = "query_gbif",
+  inaturalist = "query_inaturalist",
+  `macaulay library` = "query_macaulay",
+  observation = "query_observation",
+  wikiaves = "query_wikiaves",
+  `xeno-canto` = "query_xenocanto"
+)
+
+Function <- Function[match(names(metadata_list), names(Function))]
+
+
+file_types <- vapply(
+  Function,
+  function(x) {
+    paste(formals(get(x))$format, collapse = ", ")
+  },
+  FUN.VALUE = character(1)
+)
+file_types <- gsub("c, ", "", file_types)
+file_types[Function == "query_xenocanto"] <- "sound"
+
+file_types <- gsub("getOption,|suwo_format, c", "", file_types)
+
+# remove c(, ) and " from file_types
+file_types <- gsub("\\(|\\)|\"", "", file_types)
+
+urls <- c(
+  gbif = "https://www.gbif.org/",
+  inaturalist = "https://www.inaturalist.org/",
[…]
