### 2026-02-10 18435f8
fixing RCMD check issues
