+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: suwo
 Type: Package
 Title: Access Nature Media Repositories
-Version: 0.1.0
+Version: 0.2.0
 Maintainer: Marcelo Araya-Salas <marcelo.araya@ucr.ac.cr>
 Description: Streamline searching/downloading of nature media files (e.g. audios, photos) from online repositories. The package offers functions for obtaining media metadata from online repositories, downloading associated media files and updating data sets with new records.
 URL: https://marce10.github.io/suwo/, https://github.com/maRce10/suwo
