### 2026-02-11 74bbc23
fix query_wikiaves by waiting 1 s between request
