### 2026-02-13 f2d5f7a
remove unnecessary clusters for windows
