### 2026-02-16 1f3879e
mc.cores to suwo_cores
