+++ R/query_xenocanto.R
@@ -67,7 +67,7 @@
 query_xenocanto <-
   function(
     species = getOption("suwo_species"),
-    cores = getOption("mc.cores", 1),
+    cores = getOption("suwo_cores", 1),
     pb = getOption("suwo_pb", TRUE),
     verbose = getOption("suwo_verbose", TRUE),
     all_data = getOption("suwo_all_data", FALSE),
