+++ R/zzz.R
@@ -5,11 +5,11 @@
 
   # Package-specific default options
   op.yourpackage <- list(
-    mc.cores = 1,
-    pb = TRUE,
-    verbose = TRUE,
-    all_data = FALSE,
-    raw_data = FALSE
+    suwo_cores = 1,
+    suwo_pb = TRUE,
+    suwo_verbose = TRUE,
+    suwo_all_data = FALSE,
+    suwo_raw_data = FALSE
   )
 
   # Only set options that haven't been set by user
@@ -29,7 +29,7 @@
   original_op <- get(".original_options", envir = parent.env(environment()))
 
   # Restore only the options that were modified by our package
-  package_options <- c("mc.cores", "pb", "verbose", "all_data", "raw_data")
+  package_options <- c("suwo_cores", "suwo_pb", "suwo_verbose", "suwo_all_data", "suwo_raw_data")
 
   for (opt in package_options) {
     if (opt %in% names(original_op)) {
