### 2026-02-16 074fb68
fix typo in function name when checking arguments
