+++ DESCRIPTION
@@ -4,7 +4,7 @@ Title: Access Nature Media Repositories
 Version: 0.2.0
 Maintainer: Marcelo Araya-Salas <marcelo.araya@ucr.ac.cr>
 Description: Streamline searching/downloading of nature media files (e.g. audios, photos) from online repositories. The package offers functions for obtaining media metadata from online repositories, downloading associated media files and updating data sets with new records.
-URL: https://marce10.github.io/suwo/, https://github.com/maRce10/suwo
+URL: https://marce10.github.io/suwo/
 BugReports: https://github.com/maRce10/suwo/issues/
 License: GPL (>= 2)
 Encoding: UTF-8
@@ -32,7 +32,7 @@ Suggests:
 Depends:
     R (>= 4.0.0)
 Config/testthat/edition: 3
-RoxygenNote: 7.3.2
+RoxygenNote: 7.3.3
 Authors@R: 
     c(person("Marcelo", "Araya-Salas", 
       role = c("aut", "cre"), 
