### 2026-02-16 9167e1d
mc.cores modified to suwo_cores
