+++ R/query_wikiaves.R
@@ -202,7 +202,7 @@ query_wikiaves <-
     # loop over pages
     query_output_list <- .pbapply_sw(
       X = seq_len(nrow(id_by_page_df)),
-      cl = cl,
+      cl = cores,
       pbar = pb,
       function(x, Y = seq_len(nrow(id_by_page_df))) {
         # set index to get the right offset
