+++ man/query_xenocanto.Rd
@@ -6,7 +6,7 @@
 \usage{
 query_xenocanto(
   species = getOption("suwo_species"),
-  cores = getOption("mc.cores", 1),
+  cores = getOption("suwo_cores", 1),
   pb = getOption("suwo_pb", TRUE),
   verbose = getOption("suwo_verbose", TRUE),
   all_data = getOption("suwo_all_data", FALSE),
