+++ man/remove_duplicates.Rd
@@ -7,7 +7,7 @@
 remove_duplicates(
   metadata,
   same_repo = FALSE,
-  cores = getOption("mc.cores", 1),
+  cores = getOption("suwo_cores", 1),
   pb = getOption("suwo_pb", TRUE),
   repo_priority = c("Xeno-Canto", "GBIF", "iNaturalist", "Macaulay Library", "WikiAves",
     "Observation"),
