+++ man/query_gbif.Rd
@@ -7,7 +7,7 @@
 query_gbif(
   species = getOption("suwo_species"),
   format = getOption("suwo_format", c("image", "sound", "video", "interactive resource")),
-  cores = getOption("mc.cores", 1),
+  cores = getOption("suwo_cores", 1),
   pb = getOption("suwo_pb", TRUE),
   verbose = getOption("suwo_verbose", TRUE),
   dataset = NULL,
