### 2026-02-16 a6852e2
Update testing-checking suwo functions.Rmd
