### 2026-02-16 #729 @maRce10
Hi @Bisaloo !

thanks a lot again for your great suggestions. Here is the answers:

The logo is pkgdown/favicon/ seems out of date. You may need to run pkgdown::build_favicons(overwrite = TRUE)

**RESPONSE**: Done ([72bc7b5](https://github.com/maRce10/suwo/commit/72bc7b5254a419a49cf1c03eb939821e4a70568c))

Is this still necessary or is it handled by .pbapply_sw() now?

https://github.com/maRce10/suwo/blob/74bbc2330a03df7451b20dab8c43a05043a4c9b5/R/query_xenocanto.R#L153-L157
https://github.com/maRce10/suwo/blob/74bbc2330a03df7451b20dab8c43a05043a4c9b5/R/remove_duplicates.R#L106-L111

**RESPONSE**: Not necessary. I removed them from all functions ([f2d5f7a](https://github.com/maRce10/suwo/commit/f2d5f7a28e7ae57ae869e1325f80f8254e2b97f9))

As far as I can tell, the setting of the mc.cores option is a no-op. It won't do anything is the user set it (or set the MC_CORES env var) and if the user didn't set it, they will use one core by default. I would drop it entirely.
 The name of the other options should be updated to use the suwo_ prefix in zzz.R

**RESPONSE**: We added all “suwo_” to all options. We decide to changed “mc.cores” to “suwo_cores” as we still think it is useful to set parallelization across suwo functions in an R session. However, if the reviewer still thinks this is not a good idea we would be happy to change it.

Typo in the function name at[ https://github.com/maRce10/suwo/blob/74bbc2330a03df7451b20dab8c43a05043a4c9b5/R/find_duplicates.R#L80](https://github.com/maRce10/suwo/blob/74bbc2330a03df7451b20dab8c43a05043a4c9b5/R/find_duplicates.R#L80)

**RESPONSE**: Fixed ([074fb68](https://github.com/maRce10/suwo/commit/074fb685e0a61e8f727f0020cda913cfb5bd76a2))

. See also this item from my previous comment:
The CITATION file can easily get out of date. I would recommend:
relying on the default generated citation (when no CITATION file is present) or
using special meta fields to automatically fetch version and year from DESCRIPTION, as done in[ https://github.com/funecology/fundiversity/blob/main/inst/CITATION](https://github.com/funecology/fundiversity/blob/main/inst/CITATION).

**RESPONSE**: fixed ([b7a63c9](https://github.com/maRce10/suwo/commit/b7a63c976f334aa1fad40e5bff88edf3a591ea4c))
