### 2026-02-16 ab95d3c
update citation in README
