+++ README.Rmd
@@ -97,5 +97,9 @@ The [suwo](https://marce10.github.io/suwo/) package is designed exclusively for 
 
 Please cite [suwo](https://marce10.github.io/suwo/) as follows:
 
-Araya-Salas, M., J. Elizondo-Calvo & A. Rico-Guevara. 2025. suwo: access nature media
-      repositories. R package version 0.1.0.
+
+```{r citation, echo = FALSE}
+
+cat(format(citation("suwo"), style = "text"))
+
+```
