+++ man/suwo-package.Rd
@@ -32,7 +32,6 @@ License: GPL (>= 2)
 Useful links:
 \itemize{
   \item \url{https://marce10.github.io/suwo/}
-  \item \url{https://github.com/maRce10/suwo}
   \item Report bugs at \url{https://github.com/maRce10/suwo/issues/}
 }
 
