### 2026-02-20 544193b
fixing query_wikiaves tests
