### 2026-03-19 db5a589
more detail on error messages
