+++ reference/update_metadata.md
@@ -8,7 +8,7 @@
 update_metadata(
   metadata,
   path = ".",
-  cores = getOption("mc.cores", 1),
+  cores = getOption("suwo_cores", 1),
   pb = getOption("suwo_pb", TRUE),
   verbose = getOption("suwo_verbose", TRUE),
   api_key = NULL,
@@ -21,12 +21,12 @@ update_metadata(
 - metadata:
 
   data frame previously obtained from any suwo query function (i.e.
-  \`query_reponame()\`).
+  `query_reponame()`).
 
 - path:
 
   Directory path where the .csv file will be saved. Only applicable for
-  [`query_macaulay`](https://marce10.github.io/suwo/reference/query_macaulay.md)
+  [`query_macaulay()`](https://marce10.github.io/suwo/reference/query_macaulay.md)
   query results. By default it is saved into the current working
   directory (`"."`).
 
@@ -44,7 +44,8 @@ update_metadata(
 
   Logical argument to control if progress bar is shown. Default is
   `TRUE`. Can be set globally for the current R session via the
-  "suwo_pb" option ( `options(suwo_pb = TRUE)`).
+  "suwo_pb" option ( `options(suwo_pb = TRUE)`). Not shown if only a few
+  observations are found.
 
 - verbose:
 
@@ -58,7 +59,7 @@ update_metadata(
   authorization for searches. Get yours at
   <https://xeno-canto.org/account>. Only needed if the input metadata
   comes from
-  [`query_xenocanto`](https://marce10.github.io/suwo/reference/query_xenocanto.md).
+  [`query_xenocanto()`](https://marce10.github.io/suwo/reference/query_xenocanto.md).
 
 - dates:
 
@@ -70,7 +71,7 @@ update_metadata(
   and between 2021 to 2025 use `dates = c(2010, 2020, 2025)`. If years
   contain decimals searches will be split by months within years as
   well. Only needed if the input metadata comes from
-  [`query_macaulay`](https://marce10.github.io/suwo/reference/query_macaulay.md).
+  [`query_macaulay()`](https://marce10.github.io/suwo/reference/query_macaulay.md).
 
 ## Value
 
@@ -81,15 +82,15 @@ appended.
 
 This function updates the metadata from a previous query to add entries
 found in the source repository. All observations must belong to the same
-repository. The function adds the column \`new_entry\` which labels
-those entries that are new (i.e., not present in the input metadata).
-The input data frame must have been obtained from any of the query
-functions with the argument \`raw_data = FALSE\`. The function uses the
-same query species and format as in the original query. If no new
-entries are found, the function returns the original metadata and prints
-a message. If some old entries are not returned in the new query they
-are still retained. The function assumes that no new files are added to
-existing repository entries.
+repository. The function adds the column `new_entry` which labels those
+entries that are new (i.e., not present in the input metadata). The
+input data frame must have been obtained from any of the query functions
+with the argument `raw_data = FALSE`. The function uses the same query
+species and format as in the original query. If no new entries are
+found, the function returns the original metadata and prints a message.
+If some old entries are not returned in the new query they are still
+retained. The function assumes that no new files are added to existing
+repository entries.
 
 ## Author
 
@@ -100,7 +101,7 @@ Marcelo Araya-Salas (<marcelo.araya@ucr.ac.cr>)
 ``` r
 # query metadata
 a_gioiosa <- query_gbif(species = "Amanita gioiosa", format =  "image")
-#> ✔ Obtaining metadata (16 matching records found) 🎊:
[…]
