### 2026-03-19 0b84652
update links and workflows to work with rOpenSci
