+++ vignettes/suwo.Rmd
@@ -30,13 +30,13 @@ params:
 
 ::: {.alert .alert-info}
 
-The [suwo](https://marce10.github.io/suwo/) package aims to simplify the
+The [suwo](https://docs.ropensci.org/suwo/) package aims to simplify the
 retrieval of nature media (mostly photos, audio files and videos) across
 multiple online biodiversity databases. This vignette provides an overview of
 the package’s core querying functions, the searching and downloading of media
 files, and the compilation of metadata from various sources. For detailed
 information on each function, please refer to the
-[function reference](https://maRce10.github.io/suwo/reference/index.html)
+[function reference](https://docs.ropensci.org/suwo/reference/index.html)
 or use the help files within R (e.g., `?query_gbif`).
 
 :::
@@ -58,7 +58,7 @@ Install the latest development version from GitHub:
 
 ``` r
 # install package
-remotes::install_github("maRce10/suwo")
+remotes::install_github("ropensci/suwo")
 
 #load packages
 library(suwo)
@@ -66,7 +66,7 @@ library(suwo)
 
 # Basic workflow for obtaining nature media files
 
-Obtaining nature media using [suwo](https://marce10.github.io/suwo/) follows a basic
+Obtaining nature media using [suwo](https://docs.ropensci.org/suwo/) follows a basic
 sequence. The following diagram illustrates this workflow and the main functions
 involved:
 
@@ -81,28 +81,28 @@ Here is a description of each step:
 query functions (`query_repo_name()`) that connect to five different online
 repositories (Xeno-Canto, Inaturalist, GBIF, Macaulay Library and WikiAves). The output of these queries is a data frame containing metadata
 associated with the media files (e.g., species name, date, location, etc, see below).
-* Users can update their datasets with new records using the [update_metadata()](https://marce10.github.io/suwo/reference/update_metadata.html) function.
+* Users can update their datasets with new records using the [update_metadata()](https://docs.ropensci.org/suwo/reference/update_metadata.html) function.
 
 2. Curate metadata:
 
 * If multiple repositories are queried, the resulting metadata data frames can
 be merged into a single data frame using the
-[merge_metadata()](https://marce10.github.io/suwo/reference/merge_metadata.html) function.
+[merge_metadata()](https://docs.ropensci.org/suwo/reference/merge_metadata.html) function.
 
- * Check for duplicate records in their datasets using the [find_duplicates()](https://marce10.github.io/suwo/reference/find_duplicates.html) function. Candidate duplicated entries are identified based on matching species name, country, date, user name, and
+ * Check for duplicate records in their datasets using the [find_duplicates()](https://docs.ropensci.org/suwo/reference/find_duplicates.html) function. Candidate duplicated entries are identified based on matching species name, country, date, user name, and
 geographic coordinates. User can double check the candidate duplicates and
-decide which records to keep, which can be done with [remove_duplicates()](https://marce10.github.io/suwo/reference/remove_duplicates.html).
+decide which records to keep, which can be done with [remove_duplicates()](https://docs.ropensci.org/suwo/reference/remove_duplicates.html).
 
 3. Get media:
 
-* Download the media files associated with the metadata using the [download_media()](https://marce10.github.io/suwo/reference/download_media.html) function.
+* Download the media files associated with the metadata using the [download_media()](https://docs.ropensci.org/suwo/reference/download_media.html) function.
 
 
 
 
 # Obtaining metadata: the query functions
 
[…]
