+++ README.Rmd
@@ -14,14 +14,14 @@ editor_options:
 <!-- [![minimal R version](https://img.shields.io/badge/R%3E%3D-`r strsplit(gsub("depends: R \\(|\\)", "", grep("^DEPENDS", ignore.case = TRUE, readLines(con = "./DESCRIPTION"), value = TRUE), ignore.case = TRUE), ",")[[1]][1]`-6666ff.svg)](https://cran.r-project.org/)  -->
 <!-- [![packageversion](https://img.shields.io/badge/Package%20version-`r gsub("version: |version:", "", grep("^version", ignore.case = TRUE, readLines(con = "./DESCRIPTION"), value = TRUE), ignore.case = TRUE)`-orange.svg?style=flat-square)](commits/develop)  -->
 [![Last-changedate](https://img.shields.io/badge/last%20change-`r gsub('-', '--', Sys.Date())`-yellowgreen.svg)](/commits/main) 
-  [![Codecov test coverage](https://codecov.io/gh/maRce10/suwo/branch/main/graph/badge.svg)](https://app.codecov.io/gh/maRce10/suwo?branch=main)
-[![pkgcheck](https://github.com/maRce10/suwo/workflows/pkgcheck/badge.svg)](https://github.com/maRce10/suwo/actions?query=workflow%3Apkgcheck)  
+  [![Codecov test coverage](https://codecov.io/gh/ropensci/suwo/branch/main/graph/badge.svg)](https://app.codecov.io/gh/ropensci/suwo?branch=main)
+[![pkgcheck](https://github.com/ropensci/suwo/workflows/pkgcheck/badge.svg)](https://github.com/ropensci/suwo/actions?query=workflow%3Apkgcheck)  
 <!-- [![CRAN_Status_Badge](https://www.r-pkg.org/badges/version/suwo)](https://cran.r-project.org/package=suwo) -->
 <!-- [![Total Downloads](https://cranlogs.r-pkg.org/badges/grand-total/suwo)](https://cranlogs.r-pkg.org/badges/grand-total/suwo) -->
-[![R-CMD-check](https://github.com/maRce10/suwo/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/maRce10/suwo/actions/workflows/R-CMD-check.yaml)
+[![R-CMD-check](https://github.com/ropensci/suwo/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/ropensci/suwo/actions/workflows/R-CMD-check.yaml)
 <!-- badges: end -->
   
-The [suwo](https://marce10.github.io/suwo/) package aims to simplify the 
+The [suwo](https://docs.ropensci.org/suwo/) package aims to simplify the 
 retrieval of nature media (mostly photos, audio files and videos) across 
 multiple online biodiversity databases. The five major media repositories 
 accessed by this package (GBIF, iNaturalist, Macaulay Library, WikiAves, 
@@ -63,12 +63,14 @@ install.packages('suwo', repos = c('https://marce10.r-universe.dev'))
 library(suwo)
 ```
 To install the latest developmental version from
-[github](https://github.com/) you will need the R package
-[remotes](https://cran.r-project.org/package=remotes):
+[github](https://github.com/) run:
 
 ```{r, eval = FALSE}
-# install remotes if you don't have it yet
-remotes::install_github("maRce10/suwo")
+
+install.packages("suwo", repos = c(
+  'https://ropensci.r-universe.dev',
+  'https://cloud.r-project.org'
+))
 
 #load package
 library(suwo)
@@ -76,21 +78,21 @@ library(suwo)
 
 # Basic workflow for obtaining nature media files
 
[…]
