+++ man/suwo-package.Rd
@@ -31,7 +31,8 @@ License: GPL (>= 2)
 \seealso{
 Useful links:
 \itemize{
-  \item \url{https://marce10.github.io/suwo/}
+  \item \url{https://docs.ropensci.org/suwo}
+  \item \url{https://github.com/ropensci/suwo}
   \item Report bugs at \url{https://github.com/maRce10/suwo/issues/}
 }
 
