+++ CITATION.cff
@@ -2,13 +2,16 @@
 # CITATION file created with {cffr} R package
 # See also: https://docs.ropensci.org/cffr/
 # --------------------------------------------
-
+ 
 cff-version: 1.2.0
 message: 'To cite package "suwo" in publications use:'
 type: software
 license: GPL-2.0-or-later
 title: 'suwo: Access Nature Media Repositories'
 version: 0.2.0
+identifiers:
+- type: url
+  value: https://github.com/ropensci/suwo
 abstract: Streamline searching/downloading of nature media files (e.g. audios, photos)
   from online repositories. The package offers functions for obtaining media metadata
   from online repositories, downloading associated media files and updating data sets
@@ -25,8 +28,8 @@ authors:
   given-names: Alejandro
   email: colibri@uw.edu
   orcid: https://orcid.org/0000-0003-4067-5312
-repository-code: https://github.com/ropensci/suwo/issues/
-url: https://marce10.github.io/suwo/
+repository-code: https://github.com/maRce10/suwo/issues/
+url: https://docs.ropensci.org/suwo
 contact:
 - family-names: Araya-Salas
   given-names: Marcelo
