### 2026-03-25 fc83b2c
fix map_locations popup and added observation_url basic column
