+++ R/query_gbif.R
@@ -220,6 +220,11 @@ query_gbif <-
 
     query_output_df$gbifid <- query_output_df$scientificName <- NULL
 
+    query_output_df$observation_url <- paste0(
+      "https://www.gbif.org/occurrence/",
+      query_output_df$key
+    )
+
     query_output_df <- .format_query_output(
       X = query_output_df,
       call = base::match.call(),
