### 2026-03-26 21cce09
Update CITATION.cff
