### 2026-03-26 3ab66ea
updates on map location arguments
