+++ R/map_locations.R
@@ -6,14 +6,25 @@
 #' @param cluster Logical to control if icons are clustered by locality.
 #' Default is `FALSE`.
 #' @param marker_color Character vector  with the color(s) to be used for the
-#' markers. Possible values are "red", "darkred", "lightred", "orange",
+#' markers (when \code{type = "markers"}. Possible values are "red", "darkred", "lightred", "orange",
 #' "beige", "green", "darkgreen", "lightgreen", "blue", "darkblue",
 #' "lightblue", "purple", "darkpurple", "pink", "cadetblue", "white", "gray",
 #' "lightgray", "black". By default it "orange". Can be used to indicate the
 #' levels of a character or factor column with the argument "by". In such
 #' a case users must supplied as many colors as levels in the column.
-#' @param by Name of column to be used for coloring markers. Default is
-#' "species".
+#' @param by Character string indicating the name of the column used to group
+#' observations for coloring. Default is `"species"`. For `type = "circles"`,
+#' this determines the color mapping and legend. For `type = "markers"`,
+#' this determines how marker colors are assigned.
+#' @param type Character string indicating how observations are displayed.
+#' Options are `"circles"` (default) or `"markers"`. Circles use a color
+#' palette and include a legend, while markers use colored icons and can be
+#' clustered.
+#' @param palette Function used to generate colors for circle markers when
+#' `type = "circles"`. The function must take a single integer (`n`) and
+#' return `n` colors. By default it uses
+#' `function(n) grDevices::hcl.colors(n, "mako")`. Ignored when
+#' `type = "markers"`.
 #' @return An interacrive map with the locations of the observations.
 #' @export
 #' @name map_locations
@@ -41,29 +52,41 @@
 #' map_locations(e_hochs)
 #' }
 #' }
-#' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr}) and Grace
-#' Smith Vidaurre
+#' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
 
 map_locations <- function(
-  metadata,
-  cluster = FALSE,
-  marker_color = "orange",
-  by = "species"
+    metadata,
+    cluster = FALSE,
+    marker_color = NULL,
+    by = "species",
+    type = "circles",
+    palette = function(n) grDevices::hcl.colors(n, "mako")
 ) {
 
-  ##  argument checking
+  # argument checking
   check_results <- .check_arguments(
     fun = "map_locations",
     args = list(
-     metadata = metadata,
-     cluster = cluster,
-     marker_color = marker_color,
-     by = by
+      metadata = metadata,
+      cluster = cluster,
+      marker_color = marker_color,
+      by = by,
+      type = type,
+      palette = palette
     )
   )
 
   .report_assertions(check_results)
 
+  # validate type
+  if (!type %in% c("circles", "markers")) {
+    stop("'type' must be either 'circles' or 'markers'")
+  }
+
+  # validate palette
+  if (!is.function(palette)) {
+    stop("'palette' must be a function")
+  }
 
   # remove observations with no lat lon data
   inx_with_coors <- !is.na(metadata$latitude) |
@@ -76,50 +99,35 @@ map_locations <- function(
     )
     return(invisible(NULL))
   }
+
   metadata <- metadata[inx_with_coors, , drop = FALSE]
 
-  # make map
+  # labels for coloring
+  metadata$labels <- metadata[, by, drop = TRUE]
+  levs <- unique(metadata$labels)
 
[…]
