### 2026-04-08 4cae505
enhance map_locations() and fixes required by CRAN
