+++ R/query_macaulay.R
@@ -216,18 +216,21 @@ query_macaulay <-
       }
 
       # let users know where to save the file
-      cat(
+      .message(
         paste(
           "A browser will open the macaulay library website.",
-          "Save the .csv file ('export' button) to this directory: "
-        )
+          "Save the .csv file ('export' button) to this directory: \n",
+          paste(normalizePath(path), "/", sep = "")
+        ),
+        as = "message"
       )
-      cat(normalizePath(path), "/", sep = "")
-      cat(
+      # .message(paste(normalizePath(path), "/", sep = ""), as = "message")
+      .message(
         paste0(
           "\n(R is monitoring for new CSV files.",
           " Press ESC to stop the function)\n"
-        )
+        ),
+        as = "message"
       )
 
       # Apply to all elements
@@ -247,7 +250,7 @@ query_macaulay <-
           # extract date range to let users know while batching
           date_range <- if (
             date_ranges_df$start_month[i] == 1 &&
-              date_ranges_df$end_month[i] == 12
+            date_ranges_df$end_month[i] == 12
           ) {
             if (date_ranges_df$start_year[i] != date_ranges_df$end_year[i]) {
               paste0(
@@ -311,7 +314,7 @@ query_macaulay <-
 
           if (
             date_ranges_df$start_month[i] != 1 ||
-              date_ranges_df$end_month[i] != 12
+            date_ranges_df$end_month[i] != 12
           ) {
             search_url <- paste0(
               search_url,
@@ -331,9 +334,8 @@ query_macaulay <-
           .monitor_new_files(path = path)
 
         # let users know the name of the csv file that was read
-        cat("File:", suffix = " ")
-
-        cat(paste(new_csv_file_list[i], "\n"))
+        .message(paste("File:  \n",new_csv_file_list[i], "\n"),
+                 as = "message")
       }
     } else {
       # warning if species was also supplied
