+++ R/query_wikiaves.R
@@ -21,7 +21,7 @@
 #' birds, where users contribute georeferenced photographs and sound
 #' recordings, which are then used to build a vast database for research
 #' and conservation.
-#' @examples  \dontrun{
+#' @examples  \donttest{
 #' # search
 #' p_nattereri <- query_wikiaves(species = "Phaethornis nattereri",
 #'     format = "image")
