+++ R/map_locations.R
@@ -6,10 +6,10 @@
 #' @param cluster Logical to control if icons are clustered by locality.
 #' Default is `FALSE`.
 #' @param marker_color Character vector  with the color(s) to be used for the
-#' markers (when \code{type = "markers"}. Possible values are "red", "darkred", "lightred", "orange",
-#' "beige", "green", "darkgreen", "lightgreen", "blue", "darkblue",
-#' "lightblue", "purple", "darkpurple", "pink", "cadetblue", "white", "gray",
-#' "lightgray", "black". By default it "orange". Can be used to indicate the
+#' markers (when \code{type = "markers"}. Possible values are "red", "darkred",
+#' "lightred", "orange", "beige", "green", "darkgreen", "lightgreen", "blue",
+#' "darkblue","lightblue", "purple", "darkpurple", "pink", "cadetblue",
+#' "white", "gray", "lightgray", "black". Can be used to indicate the
 #' levels of a character or factor column with the argument "by". In such
 #' a case users must supplied as many colors as levels in the column.
 #' @param by Character string indicating the name of the column used to group
@@ -25,7 +25,17 @@
 #' return `n` colors. By default it uses
 #' `function(n) grDevices::hcl.colors(n, "mako")`. Ignored when
 #' `type = "markers"`.
-#' @return An interacrive map with the locations of the observations.
+#' @param tags Character vector with the names of the columns in
+#' \code{metadata} to be shown in the popup. Default is
+#' \code{c("repository", "key", "species", "date", "country", "locality",
+#' "user_name")}.
+#' @param show_media Logical indicating whether to display media files (audio,
+#' images, videos) in the popup windows. Default is `TRUE`.
+#' @param popup_size Numeric value that controls the size of the popups.
+#' Default is `1`. Values greater than `1` will increase the size of the
+#' popups, while values between `0` and `1` will decrease it.
+#' @return An  a `leaflet` interactive map object with the locations of
+#' the observations.
 #' @export
 #' @name map_locations
 #' @details The function uses the `leaflet` package to create interactive
@@ -34,36 +44,48 @@
 #' observation the function displays a marker in the map with a popup that
 #' shows the species name, country, locality, user name, and repository.
 #' The popup also includes an audio player for sound recordings, an image
-#' for photos and a video player for videos. Users can zoom in and out of the
-#' map and click on the markers to see the popups. If `cluster = TRUE`,
-#' markers that are close together will be clustered into a single marker
-#' that shows the number of observations in that cluster. Users can click on
-#' the cluster marker to zoom in and see the individual markers. This function
-#' is useful for exploring the geographic distribution of media records and
-#' identifying patterns or gaps in the data.
-#' @examples \dontrun{
+#' for photos and a video player for videos.
+#'
+#' When multiple media files are associated with the same observation (i.e.,
+#' identical values in the \code{key} column), they are grouped into a single
+#' popup. The first media item is displayed by default, and users can navigate
+#' through additional media using arrow buttons within the popup.
+#'
+#' Users can zoom in and out of the map and click on the markers to see the
+#' popups. If `cluster = TRUE`, markers that are close together will be
[…]
