+++ R/find_duplicates.R
@@ -45,7 +45,7 @@
 #' upload the same observation to different platforms. This can also occur as
 #' some repositories automatically share data with other repositories,
 #' particularly with GBIF.
-#' @examples \dontrun{
+#' @examples \donttest{
 #' # get metadata from 2 repos
 #' gb <- query_gbif(species = "Turdus rufiventris", format =  "sound")
 #' inat <- query_inaturalist(species = "Turdus rufiventris",
