### 2026-04-08 cf1b6a2
Update CITATION.cff
