+++ CITATION.cff
@@ -37,7 +37,7 @@ preferred-citation:
     email: marcelo.araya@ucr.ac.cr
     orcid: https://orcid.org/0000-0003-3594-619X
   notes: R package version 0.2.0
-  url: https://docs.ropensci.org/suwo/
+  url: https://docs.ropensci.org/suwo
 repository-code: https://github.com/ropensci/suwo/issues/
 url: https://docs.ropensci.org/suwo/
 contact:
@@ -320,6 +320,30 @@ references:
     orcid: https://orcid.org/0000-0002-3386-6076
   year: '2026'
   doi: 10.32614/CRAN.package.kableExtra
+- type: software
+  title: htmlwidgets
+  abstract: 'htmlwidgets: HTML Widgets for R'
+  notes: Suggests
+  url: https://github.com/ramnathv/htmlwidgets
+  repository: https://CRAN.R-project.org/package=htmlwidgets
+  authors:
+  - family-names: Vaidyanathan
+    given-names: Ramnath
+  - family-names: Xie
+    given-names: Yihui
+  - family-names: Allaire
+    given-names: JJ
+  - family-names: Cheng
+    given-names: Joe
+    email: joe@posit.co
+  - family-names: Sievert
+    given-names: Carson
+    email: carson@posit.co
+    orcid: https://orcid.org/0000-0002-4958-2844
+  - family-names: Russell
+    given-names: Kenton
+  year: '2026'
+  doi: 10.32614/CRAN.package.htmlwidgets
 - type: software
   title: 'R: A Language and Environment for Statistical Computing'
   notes: Depends
