+++ DESCRIPTION
@@ -3,7 +3,7 @@ Type: Package
 Title: Access Nature Media Repositories
 Version: 0.2.0
 Maintainer: Marcelo Araya-Salas <marcelo.araya@ucr.ac.cr>
-Description: Streamline searching/downloading of nature media files (e.g. audios, photos) from online repositories. The package offers functions for obtaining media metadata from online repositories, downloading associated media files and updating data sets with new records.
+Description: Streamline searching, downloading and formatting of nature media files (e.g. audios, photos) from online repositories. The package offers functions for obtaining media metadata from online repositories, downloading associated media files and updating data sets with new records.
 URL: https://docs.ropensci.org/suwo/, https://github.com/ropensci/suwo/
 BugReports: https://github.com/ropensci/suwo/issues/
 License: GPL (>= 2)
