### 2026-04-10 62ada89
fix for CRAN and pumping map_locations
