+++ R/internal_functions.R
@@ -157,11 +157,11 @@
 }
 
 .message <- function(
-    text = "Obtaining metadata ({n} matching record{?s} found)",
-    as = c("success", "warning", "failure", "error", "message"),
-    n = NULL,
-    suffix = "\n",
-    nfiles = NULL
+  text = "Obtaining metadata ({n} matching record{?s} found)",
+  as = c("success", "warning", "failure", "error", "message"),
+  n = NULL,
+  suffix = "\n",
+  nfiles = NULL
 ) {
   if (!is.null(n)) {
     text <- cli::pluralize(text)
@@ -219,6 +219,35 @@
   return(sanitized)
 }
 
+# build_download_url
+.build_download_url <- function(repo, key, file_url, file_extension = NULL) {
+  switch(
+    repo,
+
+    "Xeno-Canto" = {
+      key_clean <- gsub("XC", "", key)
+      paste0("https://xeno-canto.org/", key_clean, "/download")
+    },
+
+    "Macaulay Library" = {
+      paste0(
+        "https://cdn.download.ams.birds.cornell.edu/api/v1/asset/",
+        key
+      )
+    },
+
+    "iNaturalist" = {
+      sub("(/square|/small|/medium|/large)", "/original", file_url)
+    },
+
+    "WikiAves" = file_url,
+    "GBIF" = file_url,
+    "Observation" = file_url,
+
+    file_url
+  )
+}
+
 # Function to download file according to repository
 .download_basic <- function(metadata, x, path, overwrite, folder_by = NULL) {
   if (!is.null(folder_by)) {
@@ -226,22 +255,31 @@
     path <- normalizePath(file.path(path, folder_name))
   }
 
-  # if path does not exist create it
+  # create path if needed
   if (!dir.exists(path)) {
     dir.create(path, recursive = TRUE)
   }
 
-  # set destination file
+  # destination file
   destfile <- file.path(normalizePath(path), metadata$downloaded_file_name[x])
 
   exists <- file.exists(destfile)
   if (exists && !overwrite) {
-    (return("already there (not downloaded)"))
+    return("already there (not downloaded)")
   }
 
+  # build URL
+  url <- .build_download_url(
+    repo = metadata$repository[x],
+    key = metadata$key[x],
+    file_url = metadata$file_url[x],
+    file_extension = metadata$file_extension[x]
+  )
+
+  # first attempt
   dl_result <- try(
     utils::download.file(
-      url = as.character(metadata$file_url[x]),
+      url = url,
       destfile = destfile,
       quiet = TRUE,
       mode = "wb",
@@ -252,12 +290,12 @@
     silent = TRUE
   )
 
-  # if failed try again after wating 0.5 seconds
+  # retry if failed
   if (.is_error(dl_result)) {
     Sys.sleep(0.5)
     dl_result <- try(
       utils::download.file(
-        url = as.character(metadata$file_url[x]),
+        url = url, # reuse same URL
         destfile = destfile,
         quiet = TRUE,
         mode = "wb",
@@ -269,12 +307,12 @@
     )
   }
 
-  # if still failed then return FALSE
+  # result handling
   if (.is_error(dl_result)) {
     return("failed")
   } else {
     if (exists) {
-      (return("overwritten"))
+      return("overwritten")
     } else {
       return("saved")
     }
@@ -376,14 +414,14 @@
 
 # format query output dataframe to standardize column names
 .format_query_output <- function(
-    X,
-    column_names,
-    all_data,
-    format,
-    raw_data = FALSE,
-    call,
-    input_file = NA,
-    only_basic_columns = FALSE
+  X,
+  column_names,
+  all_data,
+  format,
[…]
