+++ R/map_locations.R
@@ -74,18 +74,24 @@
 #' }
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
 map_locations <- function(
-    metadata,
-    cluster = FALSE,
-    marker_color = NULL,
-    by = "species",
-    type = "circles",
-    palette = function(n) grDevices::hcl.colors(n, "mako"),
-    tags = c("repository", "key", "species", "date", "country", "locality",
-             "user_name"),
-    show_media = TRUE,
-    popup_size = 1
+  metadata,
+  cluster = FALSE,
+  marker_color = NULL,
+  by = "species",
+  type = "circles",
+  palette = function(n) grDevices::hcl.colors(n, "mako"),
+  tags = c(
+    "repository",
+    "key",
+    "species",
+    "date",
+    "country",
+    "locality",
+    "user_name"
+  ),
+  show_media = TRUE,
+  popup_size = 1
 ) {
-
   check_results <- .check_arguments(
     fun = "map_locations",
     args = list(
@@ -103,8 +109,11 @@ map_locations <- function(
 
   type <- rlang::arg_match(type, c("circles", "markers"))
 
-  metadata <- metadata[!is.na(metadata$latitude) & !is.na(metadata$longitude),
-                       , drop = FALSE]
+  metadata <- metadata[
+    !is.na(metadata$latitude) & !is.na(metadata$longitude),
+    ,
+    drop = FALSE
+  ]
 
   .if_na_empty <- function(x) ifelse(is.na(x), "", x)
 
@@ -122,45 +131,65 @@ map_locations <- function(
   }
 
   content <- lapply(metadata_list, function(df) {
-
     if (show_media) {
       media_html <- unname(as.character(.make_media_html(df)))
 
       slides <- paste0(
-        mapply(function(x, i) {
-          paste0(
-            "<div class='slide' style='display:",
-            ifelse(i == 1, "block", "none"),
-            ";'>", x, "</div>"
-          )
-        }, media_html, seq_along(media_html), SIMPLIFY = TRUE),
+        mapply(
+          function(x, i) {
+            paste0(
+              "<div class='slide' style='display:",
+              ifelse(i == 1, "block", "none"),
+              ";'>",
+              x,
+              "</div>"
+            )
+          },
+          media_html,
+          seq_along(media_html),
+          SIMPLIFY = TRUE
+        ),
         collapse = ""
       )
 
       if (length(media_html) == 1) {
-
         gallery <- paste0(
           "<div class='popup-gallery' style='text-align:center;'>",
           slides,
           "</div>"
         )
-
       } else {
         gallery <- paste0(
           "<div class='popup-gallery' style='text-align:center;'>",
           slides,
           paste0(
-            "<div style='margin-top:", 8 * popup_size, "px;'>",
+            "<div style='margin-top:",
+            8 * popup_size,
+            "px;'>",
 
             "<button onclick='prevSlide(this)' style='",
-            "margin-right:", 10 * popup_size, "px;",
-            "font-size:", 14 * popup_size, "px;",
-            "padding:", 4 * popup_size, "px ", 8 * popup_size, "px;",
+            "margin-right:",
+            10 * popup_size,
+            "px;",
+            "font-size:",
+            14 * popup_size,
+            "px;",
+            "padding:",
+            4 * popup_size,
+            "px ",
+            8 * popup_size,
+            "px;",
             "border:none; background:none; cursor:pointer;'>&#9664;</button>",
 
             "<button onclick='nextSlide(this)' style='",
[…]
