### 2026-04-10 19b77b6
bullet proof gbif queries
