+++ R/query_gbif.R
@@ -38,15 +38,19 @@
 #'
 
 query_gbif <-
-  function(species = getOption("suwo_species"),
-           format = getOption("suwo_format",
-                              c("image", "sound", "video", "interactive resource")),
-           cores = getOption("suwo_cores", 1),
-           pb = getOption("suwo_pb", TRUE),
-           verbose = getOption("suwo_verbose", TRUE),
-           dataset = NULL,
-           all_data = getOption("suwo_all_data", FALSE),
-           raw_data = getOption("suwo_raw_data", FALSE)) {
+  function(
+    species = getOption("suwo_species"),
+    format = getOption(
+      "suwo_format",
+      c("image", "sound", "video", "interactive resource")
+    ),
+    cores = getOption("suwo_cores", 1),
+    pb = getOption("suwo_pb", TRUE),
+    verbose = getOption("suwo_verbose", TRUE),
+    dataset = NULL,
+    all_data = getOption("suwo_all_data", FALSE),
+    raw_data = getOption("suwo_raw_data", FALSE)
+  ) {
     ##  argument checking
     check_results <- .check_arguments(
       fun = "query_gbif",
@@ -70,9 +74,10 @@ query_gbif <-
     }
 
     ##  format
-    format <- rlang::arg_match(format,
-                               values = c("image", "sound", "video",
-                                          "interactive resource"))
+    format <- rlang::arg_match(
+      format,
+      values = c("image", "sound", "video", "interactive resource")
+    )
 
     gbif_format <- switch(
       format,
@@ -85,8 +90,11 @@ query_gbif <-
     ## base request
     base_request <- httr2::request("https://api.gbif.org/v1/occurrence/search")
 
-    base_request <- httr2::req_user_agent(base_request,
-                                          "suwo (https://github.com/ropensci/suwo)")
+    base_request <-
+      httr2::req_user_agent(
+        base_request,
+        "suwo (https://github.com/ropensci/suwo)"
+      )
 
     base_request <- httr2::req_url_query(
       base_request,
@@ -98,8 +106,9 @@ query_gbif <-
 
     base_request <- httr2::req_error(
       base_request,
-      is_error = function(resp)
+      is_error = function(resp) {
         FALSE
+      }
     )
 
     ##  first request (metadata)
@@ -140,40 +149,48 @@ query_gbif <-
       FUN = function(x, Y = offsets) {
         i <- Y[x]
 
-        resp <- try(
-          httr2::req_perform(httr2::req_url_query(base_request, offset = i)),
-          silent = TRUE)
+        resp <-
+          try(
+            httr2::req_perform(httr2::req_url_query(base_request, offset = i)),
+            silent = TRUE
+          )
 
         if (.is_error(resp) || httr2::resp_is_error(resp)) {
           return(resp)
         }
 
-        query_output <- try(httr2::resp_body_json(resp, simplifyVector = TRUE),
-                            silent = TRUE)
+        query_output <- try(
+          httr2::resp_body_json(resp, simplifyVector = TRUE),
+          silent = TRUE
+        )
 
         if (.is_error(query_output)) {
           return(query_output)
         }
 
-        if (is.null(query_output$results) ||
-            nrow(query_output$results) == 0) {
+        if (
+          is.null(query_output$results) ||
+            nrow(query_output$results) == 0
+        ) {
           return(NULL)
         }
 
-        query_output$results <- lapply(seq_len(nrow(query_output$results)),
-                                       function(u) {
-          x <- query_output$results[u, ]
+        query_output$results <- lapply(
+          seq_len(nrow(query_output$results)),
+          function(u) {
+            x <- query_output$results[u, ]
 
[…]
