+++ vignettes/suwo.Rmd
@@ -3,8 +3,8 @@ title: package overview
 subtitle: "suwo: access nature media repositories"
 pagetitle: suwo package overview
 author:
-- Jorge Elizondo, <a href="https://marce10.github.io/">Marcelo Araya-Salas</a> & <a href="https://ecophysics.org/">Alejandro Rico-Guevara</a>
-date: "`r Sys.Date()`"
+- <a href="https://marce10.github.io/">Marcelo Araya-Salas</a>, Jorge Elizondo & <a href="https://ecophysics.org/">Alejandro Rico-Guevara</a>
+date:  "2026-04-13"
 output:
   rmarkdown::html_document:
     self_contained: yes
@@ -54,7 +54,18 @@ This package is designed exclusively for non-commercial, scientific purposes, in
 
 # Installation
 
+Installing from CRAN:
 
+
+``` r
+#Install from CRAN:
+
+# From CRAN would be
+install.packages("suwo")
+
+#load package
+library(suwo)
+```
 
 Install the latest development version from GitHub:
 
@@ -68,6 +79,7 @@ install.packages("suwo", repos = c(
 #load package
 library(suwo)
 ```
+
 # Basic workflow for obtaining nature media files
 
 Obtaining nature media using [suwo](https://docs.ropensci.org/suwo/) follows a basic
@@ -188,12 +200,15 @@ These are some example queries:
 
 
 ``` r
+# Load suwo package
+library(suwo)
+
 h_sarapiquensis <- query_inaturalist(species = "Heliconia sarapiquensis",
                                      format = "image")
 ```
 
 ```
-✔ Obtaining metadata (29 matching records found) 🎊
+✔ Obtaining metadata (29 matching records found) 😀
 ```
 
 ``` r
@@ -402,7 +417,7 @@ p_lotor <- query_gbif(species = "Procyon lotor", format = "video")
 ```
 
 ```
-✔ Obtaining metadata (13 matching records found) 🥇
+✔ Obtaining metadata (13 matching records found) 😸
 ```
 
 ``` r
@@ -576,13 +591,21 @@ p_striigularis <- query_macaulay(species = "Phaethornis striigularis",
 ```
 
 ```
-A browser will open the macaulay library website. Save the .csv file ('export' button) to this directory: /home/m/Dropbox/R_package_testing/suwo/vignettes/
-(R is monitoring for new CSV files. Press ESC to stop the function)
-File:  ML__2026-04-08T19-23_stther2_video.csv 
+ℹ A browser will open the macaulay library website. Save the .csv file ('export' button) to this directory: 
+/home/m/Dropbox/R_package_testing/suwo/vignettes/ 
 ```
 
 ```
-✔ 27 matching records found 😀
+ℹ (R is monitoring for new CSV files. Press ESC to stop the function)
+```
+
+```
+ℹ File:  
+ML__2026-04-13T20-24_stther2_video.csv 
+```
+
+```
+✔ 27 matching records found 🎉
 ```
 
 Users must click on the "Export" button to save the .csv file with the metadata:
@@ -604,13 +627,21 @@ jaguar <- query_macaulay(taxon_code = "t-11032765",
 ```
 
 ```
-A browser will open the macaulay library website. Save the .csv file ('export' button) to this directory: /home/m/Dropbox/R_package_testing/suwo/vignettes/
-(R is monitoring for new CSV files. Press ESC to stop the function)
-File:  ML__2026-04-08T19-24_t-11032765_video.csv 
+ℹ A browser will open the macaulay library website. Save the .csv file ('export' button) to this directory: 
[…]
