### 2026-04-14 c8f1d21
fixing navbar for ropensci
