### 2026-04-15 b2a0a2d
add full screen to map locations
