+++ R/map_locations.R
@@ -3,8 +3,17 @@
 #' `map_locations` creates maps to visualize the geographic spread of
 #' media records.
 #' @inheritParams template_params
+#' @param metadata Data frame with the metadata of the media records to be
+#' mapped. Typically the output of one of the query functions in this package
+#' (e.g. [query_gbif()], [query_inaturalist()], etc.) or metadata formatting
+#' functions (e.g. [merge_metadata()], [remove_duplicates()], etc.). Note that
+#' only observations with geographic coordinates (i.e., non-missing values in
+#' the `latitude` and `longitude` columns) are displayed in the map.
 #' @param cluster Logical to control if icons are clustered by locality.
-#' Default is `FALSE`.
+#' Default is `FALSE`. Only applies when `type = "markers"`. When
+#' `cluster = TRUE`, markers that are close together will be clustered into a
+#' single marker that shows the number of observations in that cluster. Users
+#' can click on the cluster marker to zoom in and see the individual markers.
 #' @param marker_color Character vector  with the color(s) to be used for the
 #' markers (when \code{type = "markers"}. Possible values are "red", "darkred",
 #' "lightred", "orange", "beige", "green", "darkgreen", "lightgreen", "blue",
