### 2026-04-15 e9d30e5
Update CITATION.cff
