+++ tests/testthat/test_update_metadata.R
@@ -118,7 +118,6 @@ test_that("update query_macaulay with paging", {
   sub_df <-
     df1[df1$key != names(which.max(table(df1$key))), ]
 
-
   up_df <- update_metadata(
     metadata = sub_df,
     path = tempdir(),
