+++ vignettes/references.bib
@@ -1,0 +1,32 @@
+@article{pagano2023,
+  author = {Pagano, Tiago P. and Loureiro, Rafael B. and Lisboa, Fernanda V. N. and Peixoto, Rodrigo M. and Guimarães, Guilherme A. S. and Cruz, Gustavo O. R. and Araujo, Maira M. and Santos, Lucas L. and Cruz, Marco A. S. and Oliveira, Ewerton L. S. and Winkler, Ingrid and Nascimento, Erick G. S.},
+  title = {Bias and unfairness in machine learning models: A systematic review},
+  journal = {Big Data and Cognitive Computing},
+  year = {2023},
+  volume = {7},
+  number = {1},
+  pages = {15},
+  doi = {10.3390/bdcc7010015}
+}
+
+@article{zamudio2016,
+  author = {Zamudio, Kelly R. and Bell, R. C. and Mason, N. A.},
+  title = {Phenotypes in phylogeography},
+  journal = {PNAS},
+  year = {2016},
+  volume = {113},
+  number = {29},
+  pages = {8041--8048},
+  doi = {10.1073/pnas.1602237113}
+}
+
+@article{wang2008,
+  author = {Wang, Ian J. and Shaffer, H. Bradley},
+  title = {Rapid color evolution in an aposematic species: A phylogenetic analysis of color variation in the strikingly polymorphic strawberry poison-dart frog},
+  journal = {Evolution},
+  year = {2008},
+  volume = {62},
+  number = {11},
+  pages = {2742--2759},
+  doi = {10.1111/j.1558-5646.2008.00507.x}
+}
