+++ vignettes/explore_geographic_variation.Rmd
@@ -13,6 +13,8 @@ output:
     toc_float: true
 editor_options: 
   chunk_output_type: console
+bibliography: references.bib
+link-citations: true
 ---
 
 ```{r setup, include=FALSE}
@@ -113,19 +115,19 @@ alouatta_sound <- merged_data[grep("Alouatta", merged_data$species), ]
 
 ::: {.alert .alert-info}
 
-The goal of this vignette is to illustrate how to use the `map_locations()` function to explore the geographic distribution of media records. This function provides an interactive way to visualize where different types of media (images, sounds, and videos) have been recorded across geographic space.
+The goal of this vignette is to illustrate how to use the [map_locations()](https://docs.ropensci.org/suwo/reference/map_locations.html) function to explore the geographic distribution of media records as a proxy to understand geographic variation in biological traits.
 
 :::
 
-Understanding the geographic distribution of biological variation is essential not only for ecological and evolutionary research but also for building robust automated classification models that account for phenotypic diversity. The map_locations() function in **suwo** allows researchers to interactively explore the spatial distribution of media records retrieved through the package, enabling rapid visual assessment of geographic patterns in diverse trait data.
+Understanding the geographic distribution of biological variation is essential not only for ecological and evolutionary research [@zamudio2016] but also for building robust automated classification models that account for phenotypic diversity [@pagano2023]. The [map_locations()](https://docs.ropensci.org/suwo/reference/map_locations.html) function in **suwo** allows researchers to interactively explore the spatial distribution of media records retrieved through the package, enabling rapid visual assessment of geographic patterns in diverse trait data.
 
-The function accepts a metadata data frame —typically obtained from any of the `query_()` functions or as the output of `merge_metadata()` and `remove_duplicates()`— and generates an interactive map powered by the leaflet package. Each point on the map represents a unique observation with valid geographic coordinates. Clicking on a point opens a popup window displaying the associated media file(s) and selected metadata fields.
+The function accepts a metadata data frame —typically obtained from any of the `query_()` functions or as the output of [merge_metadata()](https://docs.ropensci.org/suwo/reference/merge_metadata.html) and [remove_duplicates()](https://docs.ropensci.org/suwo/reference/remove_duplicates.html) — and generates an interactive map powered by the [leaflet package](https://CRAN.R-project.org/package=leaflet). Each point on the map represents a unique observation with valid geographic coordinates. Clicking on a point opens a popup window displaying the associated media file(s) and selected metadata fields.
 
-In this vignette, we demonstrate how to leverage the key features of `map_locations()` through three case studies: mapping image records of a color‑polymorphic frog, mapping sound recordings of the same species, and mapping video recordings of a bird's courtship display.
+In this vignette, we demonstrate how to leverage the key features of [map_locations()](https://docs.ropensci.org/suwo/reference/map_locations.html) through three case studies: mapping image records of a color‑polymorphic frog, mapping sound recordings of the same species, and mapping video recordings of a bird's courtship display.
 
 # Mapping images
 
[…]
