+++ man/remove_duplicates.Rd
@@ -9,8 +9,7 @@ remove_duplicates(
   same_repo = FALSE,
   cores = getOption("suwo_cores", 1),
   pb = getOption("suwo_pb", TRUE),
-  repo_priority = c("Xeno-Canto", "GBIF", "iNaturalist", "Macaulay Library", "WikiAves",
-    "Observation"),
+  repo_priority = c("Xeno-Canto", "GBIF", "iNaturalist", "Macaulay Library", "WikiAves"),
   verbose = getOption("suwo_verbose", TRUE)
 )
 }
@@ -43,7 +42,7 @@ are found.}
 \item{repo_priority}{Character vector indicating the priority of
 repositories when selecting which observation to retain when duplicates
 are found. Default is \code{c("Xeno-Canto", "GBIF", "iNaturalist",
-"Macaulay Library", "Wikiaves", "Observation")}, which gives priority to
+"Macaulay Library", "Wikiaves")}, which gives priority to
 repositories in which media downloading is more
 straightforward (Xeno-Canto and GBIF).}
 
