### 2026-04-16 adbb6d1
fix citation in readme
