+++ CITATION.cff
@@ -10,6 +10,8 @@ license: GPL-2.0-or-later
 title: 'suwo: Access Nature Media Repositories'
 version: 0.2.1
 identifiers:
+- type: doi
+  value: 10.32614/CRAN.package.suwo
 - type: url
   value: https://github.com/ropensci/suwo/
 abstract: Streamline searching, downloading and formatting of nature media files (e.g.
@@ -36,9 +38,17 @@ preferred-citation:
     given-names: Marcelo
     email: marcelo.araya@ucr.ac.cr
     orcid: https://orcid.org/0000-0003-3594-619X
+  - family-names: Elizondo-Calvo
+    given-names: Jorge
+    orcid: https://orcid.org/0009-0004-5873-9585
+  - family-names: Rico-Guevara
+    given-names: Alejandro
+    email: colibri@uw.edu
+    orcid: https://orcid.org/0000-0003-4067-5312
   year: '2026'
   notes: R package version 0.2.1
   url: https://docs.ropensci.org/suwo/
+repository: https://CRAN.R-project.org/package=suwo
 repository-code: https://github.com/ropensci/suwo/issues/
 url: https://docs.ropensci.org/suwo/
 contact:
