+++ NEWS.md
@@ -3,6 +3,14 @@ suwo 0.2.1
 
 ### Minor changes
 
+* Added 'file_size' column to `download_media()` output to provide users with information about the size of downloaded media files and potentially help with detecting duplicates.
+
+
+suwo 0.2.1
+=========================
+
+### Minor changes
+
 * Added images, audio player and video player to the popups of `map_locations()` for better visualization of media types.
 
 suwo 0.2.0
