+++ CITATION.cff
@@ -8,7 +8,7 @@ message: 'To cite package "suwo" in publications use:'
 type: software
 license: GPL-2.0-or-later
 title: 'suwo: Access Nature Media Repositories'
-version: 0.2.1
+version: 0.2.2
 identifiers:
 - type: doi
   value: 10.32614/CRAN.package.suwo
@@ -46,7 +46,7 @@ preferred-citation:
     email: colibri@uw.edu
     orcid: https://orcid.org/0000-0003-4067-5312
   year: '2026'
-  notes: R package version 0.2.1
+  notes: R package version 0.2.2
   url: https://docs.ropensci.org/suwo/
 repository: https://CRAN.R-project.org/package=suwo
 repository-code: https://github.com/ropensci/suwo/issues/
