### 2026-04-23 f14f543
added description file size to vignette
