+++ vignettes/suwo.Rmd
@@ -1702,6 +1702,7 @@ head(dedup_metadata[, grep("url", names(dedup_metadata), invert = TRUE)], 4)
 
 When duplicates are found, one observation from each group of duplicates is retained in the output data frame. However, if multiple observations from the same repository are labeled as duplicates, by default (`same_repo = FALSE`) all of them are retained in the output data frame. This is useful as it can be expected that observations from the same repository are not true duplicates (e.g. different recordings uploaded to Xeno-Canto with the same date, time and location by the same user), but rather have not been documented with enough precision to be told apart. This behavior can be modified. If `same_repo = TRUE`, only one of the duplicated observations from the same repository will be retained in the output data frame (and all other excluded). The function will give priority to repositories in which media downloading is more straightforward (i.e. Xeno-Canto, GBIF), but this can be modified with the argument `repo_priority`.
 
+Further identification of duplicates can be done by checking the file size of the media (ideally including it as a criterion in the 'criteria' argument). To obtain the file size, the media files need to be downloaded first with the function [download_media()](https://docs.ropensci.org/suwo/reference/download_media.html) which returns the column `file_size` which can be used for this purpose.
 
 # Download media files
 
