### 2026-04-24 #729 @Jorge-Elizondo-C
Hi @Aariq, @Bisaloo,

Marcelo and I wanted to reach out to see if you’d be interested in being included as co-authors of the package.

Let us know if you’d be open to discussing this further.
