### 2026-04-24 #729 @Aariq
I would be comfortable if you want to include me as a contributor (`role = "ctb"`), but probably not co-author (i.e. not `role = "aut"`)
