### 2026-04-25 #729 @Bisaloo
Thanks for the offer! I agree, `role = "ctb"` or `role = "rev"` seems good to me 👍
