+++ DESCRIPTION
@@ -45,5 +45,13 @@ Authors@R:
       person("Alejandro", "Rico-Guevara", 
       role = c("aut"), 
       email = "colibri@uw.edu", 
-      comment = c(ORCID = "0000-0003-4067-5312")))
+      comment = c(ORCID = "0000-0003-4067-5312")),
+      person("Hugo", "Gruson", 
+      role = c("rev"), 
+      email = "hugo.gruson@normalesup.org", 
+      comment = c(ORCID = "0000-0002-4094-1476")),
+      person("Eric", "R", "Scott", 
+      role = c("rev"), 
+      email = "ericrscott@arizona.edu", 
+      comment = c(ORCID = "0000-0002-7430-7879")))
 VignetteBuilder: knitr
