+++ .github/workflows/readme.yml
@@ -10,10 +10,10 @@ name: Render README
 
 jobs:
   render:
+    runs-on: ubuntu-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-    name: Render README
-    runs-on: ubuntu-latest
+
     steps:
       - uses: actions/checkout@v2
 
@@ -24,15 +24,23 @@ jobs:
 
       - name: Prepare repository
         run: git checkout "${GITHUB_REF:11}"
-      - uses: r-lib/actions/setup-pandoc@v2
 
+      - uses: r-lib/actions/setup-pandoc@v2
       - uses: r-lib/actions/setup-r@v2
+
       - uses: r-lib/actions/setup-r-dependencies@v2
+
+      # ✅ Install your package (this fixes citation error)
+      - name: Install package
+        run: Rscript -e 'install.packages("devtools"); devtools::install(".", upgrade = "never", dependencies = FALSE)'
+
+      # ✅ Render README
       - name: Render README
-        run: Rscript -e 'rmarkdown::render("README.Rmd", output_format = "github_document", params = list(AIRTABLE_API_KEY = ${{ secrets.AIRTABLE_API_KEY }}))'
-        env: # Set the secrets as an input
+        run: Rscript -e 'rmarkdown::render("README.Rmd", output_format = "github_document")'
+        env:
           AIRTABLE_API_KEY: ${{ secrets.AIRTABLE_API_KEY }}
           AIRTABLE_ID: ${{ secrets.AIRTABLE_ID }}
+
       - name: Commit results
         run: |
           git commit README.md -m 'Re-build README.Rmd' || echo "No changes to commit"
