+++ .github/workflows/readme.yml
@@ -32,7 +32,7 @@ jobs:
 
       # ✅ Install your package (this fixes citation error)
       - name: Install package
-        run: Rscript -e 'install.packages("devtools"); devtools::install(".", upgrade = "never", dependencies = FALSE)'
+         run: Rscript -e 'remotes::install_local(".", dependencies = FALSE)'
 
       # ✅ Render README
       - name: Render README
