### 2026-05-04 24a38cf
fix query_macaulay no detecting csv
